==> code/R/20231214_Manuscript/1_Figure/01_read_single-cell-data_0px.Rmd <==
---
title: "Read in steinbock data"
output: html_document
date: '2022-08-12'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Read in the data {#read-data}

This section describes how to read in single-cell data and images into `R`
**after** image processing and segmentation (see Section \@ref(processing)).

We use the [imcRtools](https://github.com/BodenmillerGroup/imcRtools) package to
read in single-cell data extracted using the `steinbock` framework or the IMC
Segmentation Pipeline. Both image processing approaches also generate
multi-channel images and segmentation masks that can be read into `R` using the
[cytomapper](https://github.com/BodenmillerGroup/cytomapper) package.

```{r, message=FALSE}
library(imcRtools)
#testing
library(cytomapper)
library(BiocParallel)
```

## Read in single-cell information

For single-cell data analysis in `R` the
[SingleCellExperiment](https://bioconductor.org/packages/release/bioc/html/SingleCellExperiment.html)
[@Amezquita2019] data container is commonly used within the Bioconductor
framework. It allows standardized access to (i) expression data, (ii) cellular
metadata (e.g. cell type), (iii) feature metadata (e.g. marker name) and (iv)
experiment-wide metadata. For an in-depth introduction to the `SingleCellExperiment`
container, please refer to the [SingleCellExperiment class](https://bioconductor.org/books/3.14/OSCA.intro/the-singlecellexperiment-class.html).

The [SpatialExperiment](https://bioconductor.org/packages/release/bioc/html/SingleCellExperiment.html)
class [@Righelli2021] is an extension of the `SingleCellExperiment` class. It
was developed to store spatial data in addition to single-cell data and an
extended introduction is accessible
[here](https://bioconductor.org/packages/release/bioc/vignettes/SpatialExperiment/inst/doc/SpatialExperiment.html).

To read in single-cell data generated by the `steinbock` framework or the IMC
Segmentation Pipeline, the `imcRtools` package provides the `read_steinbock` and
`read_cpout` function, respectively. By default, the data is read into a
`SpatialExperiment` object; however, data can be read in as a
`SingleCellExperiment` object by setting `return_as = "sce"`. All functions
presented in this book are applicable to both data containers.

### steinbock generated data

The [steinbock](https://github.com/BodenmillerGroup/steinbock) framework was
used to process and segment IMC raw data available.

The `read_steinbock` function provided by `imcRtools` can now be used to read in
`steinbock` generated data. For more information, please refer to
`?read_steinbock`.

```{r read-steinbock}
path <- "/mnt/Multimodal_Imaging_Daria/PIPELINE/NEWEST_RESULTS"

spe <- read_steinbock(path=path, image_file=NULL, panel_file = file.path(path, 'NB_Panel.csv'), intensities_folder = "intensities-0px-ref",regionprops_folder = "refined_regionprops", graphs_folder=NULL) 

#Rename third IF channel (CD45 for PT, CD56 for BM)
rownames(spe)[44] <- "IF3_CD45/56_mean"
rownames(spe)[88] <- "IF3_CD45/56_mean-80"

spe
```

By default, single-cell data is read in as `SpatialExperiment` object. 
The summarized pixel intensities per channel and cell (here mean intensity) are
stored in the `counts` slot. Columns represent cells and rows represent channels.

```{r counts}
rowData(spe)
counts(spe)[45:88,1:9]
```

Metadata associated to individual cells are stored in the `colData` slot. After
initial image processing, these metadata include the numeric identifier (`ObjectNumber`),
the area, and morphological features of each cell. In addition, `sample_id` stores
the image name from which each cell was extracted and the width and height of the
corresponding images are stored.

```{r colData}
head(colData(spe))
typeof(colData(spe))
class(spe)
class(colData(spe))
nrow(colData(spe))
colData(spe)[1:5, 2:ncol(colData(spe))]
```

The main difference between the `SpatialExperiment` and the
`SingleCellExperiment` data container in the current setting is the way spatial
locations of all cells are stored. For the `SingleCellExperiment` container, the
locations are stored in the `colData` slot while the `SpatialExperiment`
container stores them in the `spatialCoords` slot:

```{r spatialCoords}
head(spatialCoords(spe))
```

Finally, metadata regarding the channels are stored in the `rowData` slot. This
information is extracted from the `panel.csv` file. Channels are ordered by
isotope mass and therefore match the channel order of the multi-channel images
(see Section \@ref(read-images)).

```{r rowData}
head(rowData(spe))
rowData(spe)
```

## Single-cell processing {#cell-processing}

After reading in the single-cell data, few further processing steps need to be
taken.

**Add additional metadata**

We can set the `colnames` of the object to generate unique identifiers per cell:

```{r set-colnames}
colnames(spe) <- paste0(spe$sample_id, "_", spe$ObjectNumber)
```

It is also often the case that sample-specific metadata are available externally.
For the current data, we need to link metadata from an external excel file.
More information on how to extract matching patterns from a string can be found [here](https://stringr.tidyverse.org/reference/str_extract.html).

```{r add-indication}
library(stringr)

spe$sample <- gsub(spe$sample_id, pattern="([UM]).*", replacement="\\1")
spe$fm_id <- str_match(spe$sample, "_.*_(.*)_")[,2]  
spe$tissue <- str_match(spe$sample, "^.*_.*_(.*)")[,2] 
spe$tissue[spe$tissue=="TU"] = "PT"

meta <- read.csv(file.path(path, '20231128_metadata_complete.csv'))

spe$puncture_side <- meta$Puncture_side[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$infiltration_right <- meta$Infiltration_right[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$infiltration_left <- meta$Infiltration_left[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$exclude <- meta$exclude[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$imc_score <- meta$IMC_score[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$staining_date <- meta$Staining_date[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$stainer <- meta$Stainer[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$imc_date <- meta$IMC_date[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$DE_date <- meta$DE_date[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$DE_year <- sub("-.*", "", spe$DE_date)
spe$control <- meta$control[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$study_id <- meta$Study_ID[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$timepoint <- meta$Timepoint[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$stage <- meta$stage4[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$progression <- meta$efs[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$tod <- meta$tod[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$efs_dur <- meta$efs_dur[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$tod_date <- meta$tod_dat1[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$tod_dur <- meta$tod_dur[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$rez_date <- meta$rez_dat[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$prog_date <- meta$prog_dat[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$event_date <- meta$event_dat[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X1p_loss <- meta$X1p_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X1q_gain <- meta$X1q_gain[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X1q_loss <- meta$X1q_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X2p_gain <- meta$X2p_gain[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X3p_loss <- meta$X3p_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X3q_loss <- meta$X3q_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X4p_loss <- meta$X4p_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X6q_loss <- meta$X6q_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X7q_gain <- meta$X7q_gain[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$PTPRD_del <- meta$PTPRD_del[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$CDKN2AB_del <- meta$CDKN2AB_del[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X11q_loss <- meta$X11q_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X12q_gain <- meta$X12q_gain[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X14q_loss <- meta$X14q_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X17p_loss <- meta$X17p_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X17q_gain <- meta$X17q_gain[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X19p_loss <- meta$X19p_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X19q_loss <- meta$X19q_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X22q_loss <- meta$X22q_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$MYCN_amp <- meta$MYCN_amp[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$TERT_rear <- meta$TERT_rear[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$ATRX_del <- meta$ATRX_del[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$ALK_amp <- meta$ALK_amp[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$ALK_mut <- meta$ALK_mut[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]

spe$age_diag <- meta$age_at_diag[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$therapy <- meta$therapy[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]

spe$ROI <- str_match(spe$sample_id, "^.*_.*_.*_.*_(.*)")[,2]
spe$year <- gsub(spe$sample_id, pattern="^.*_.*_(.*)-.*", replacement="\\1")

```


```{r define-color-code, message=FALSE}
library(RColorBrewer)
library(circlize)
library(paletteer)

meta_path <- file.path(path, '20231128_metadata_complete.csv')

qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector_74 = unlist(mapply(brewer.pal, qual_col_pals$maxcolors, rownames(qual_col_pals)))

col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]

meta_all = read.csv(meta_path)
meta_all <- meta_all[order(meta_all$Study_ID),] #order by study-id

exclude <- replace(x = meta_all$exclude, list =  !meta_all$exclude %in% c('no', 'maybe'), values =  'partially')

DE_year <- sub("-.*", "", meta_all$DE_date)

metadata(spe)$color_vectors$col_sample = setNames(sample(col_vector_433, length(unique(meta_all$Sample))), unique(meta_all$Sample))
metadata(spe)$color_vectors$col_puncture_side = setNames(brewer.pal(length(unique(meta_all$Puncture_side)), 'Set3'), unique(meta_all$Puncture_side))
metadata(spe)$color_vectors$col_staining_batch = setNames(sample(col_vector_74, length(unique(meta_all$Staining_date))), unique(meta_all$Staining_date))
metadata(spe)$color_vectors$col_stainer = setNames(c('#a6611a', '#dfc27d', '#80cdc1', '#018571'), unique(meta_all$Stainer))
metadata(spe)$color_vectors$col_imc_batch = setNames(sample(col_vector_74, length(unique(meta_all$IMC_date))), unique(meta_all$IMC_date))
metadata(spe)$color_vectors$col_imc_score = colorRamp2(c(min(meta_all$IMC_score, na.rm = TRUE), max(meta_all$IMC_score, na.rm = TRUE)/2, max(meta_all$IMC_score, na.rm = TRUE)), c("red", "white", "blue")) 
metadata(spe)$color_vectors$col_study_id = setNames(sample(col_vector_433, length(unique(meta_all$Study_ID))), unique(meta_all$Study_ID))
metadata(spe)$color_vectors$col_DE_year = setNames(sample(col_vector_74, length(unique(DE_year))), unique(DE_year))
metadata(spe)$color_vectors$col_timepoint = setNames(c('#7fc97f', '#beaed4', '#fdc086' , '#ffff99', '#386cb0'), unique(meta_all$Timepoint))
metadata(spe)$color_vectors$col_aberrations = setNames(c('#018571','#a6611a','#80cdc1','#f5f5f5'), c(0,1,3,9))
metadata(spe)$color_vectors$col_stages = setNames(c('#a6cee3', '#1f78b4', '#b2df8a', '#33a02c', '#fb9a99', '#e31a1c'), unique(meta_all$stage4))
metadata(spe)$color_vectors$col_efs_dur = colorRamp2(c(min(meta_all$efs_dur, na.rm = TRUE), max(meta_all$efs_dur, na.rm = TRUE)/2, max(meta_all$efs_dur, na.rm = TRUE)), paletteer_c("ggthemes::Red-Green Diverging", 3))
metadata(spe)$color_vectors$col_tissue = setNames(c('aquamarine4','brown'), unique(meta_all$Tissue))
metadata(spe)$color_vectors$col_control = setNames(c('mediumseagreen', 'gray77'), unique(meta_all$control))
metadata(spe)$color_vectors$col_exclude = setNames(c('#018571','#a6611a','#80cdc1'), unique(exclude))
metadata(spe)$color_vectors$col_progress = setNames(c('gray77', 'palegreen3', 'firebrick1'), c('n.a.', 0, 1))
```

**Check for technical confounders (batch effects)**

```{r plot by sample}

selected_samples <- unique(spe$sample)[170:178]
marker <- 'IF3_CD45/56_mean-80'

library(dittoSeq)
dittoRidgePlot(spe[,spe$sample%in%selected_samples], var = marker, group.by = "sample", assay = "counts") +
    ggtitle(paste(marker,  "- per sample"))
```

```{r plot by staining date, fig.height=10}

#marker <- 'CD45_Eu153_mean'
marker <- 'GD2_Gd155_mean'
tissue <- unique(spe$tissue)
#tissue <- "PT"
dittoRidgePlot(spe[,spe$tissue%in%tissue], var = marker, group.by = "tissue", assay = "counts") +
    ggtitle(paste(marker,  "- per tissue"))

```


## Save objects

Finally, the generated data objects can be saved for further downstream 
processing and analysis.We will exclude low-quality samples.

```{r save-objects-read-data}

assay(spe, "exprs") <- assay(spe, "counts")

saveRDS(spe, file.path(path, "results/Manuscript/1_Figure/spe_0px_refined.rds"))
```

## Session Info

<details>
   <summary>SessionInfo</summary>
   
```{r, echo = FALSE}
sessionInfo()
```
</details>
