==> code/R/20231214_Manuscript/1_Figure/03_removal_lost_cells.Rmd <==
---
title: "Exclusion of lost (swam-off) cells"
output: html_document
date: '2022-12-28'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ggplot2)
```

In this RMD file, we will identify cells that swam off between IF and IMC and exclude those from further downstream analysis. First we will read in the single-cell data which were based on masks without expansion to make sure that we do not catch signal from neighboring cells.

```{r read-data}

path <- "/mnt/Multimodal_Imaging_Daria/PIPELINE/NEWEST_RESULTS"

spe <- readRDS(file.path(path,"results/Manuscript/1_Figure/spe_0px_refined.rds"))

```


Next, we will exclude PT cells with an area< 100px and BM cells with an area <250 and >15600, as we identified the respective cells as segmentation artifacts in 02_integrationmorph-features.Rmd. Then we will generate next generate a matrix for the spatial experiment object and exclude IF markers DAPI, GD2 and CD56. Then, we will examine whether there are still "ghosts" left by clustering the reduced matrix. We will use mean-80 instead of mean intensity, because cells might have a low mean intensity due to their size, which should not be excluded. As there are sometimes cells, without nuclear staining, like clPARP+ cells, we will include all IMC markers and DAPI for clustering. 

```{r generate-dataframes, message=FALSE, fig.height=7}
library(bruceR)

assay(spe, "exprs") <- assay(spe, "counts")
intensity_mat <- assay(spe, "exprs")
colnames(intensity_mat) <- spe$sample

#use mean-80 intensity to avoid losing big cells
intensity_mat_mean_80 <- intensity_mat[grepl("mean-80", rownames(intensity_mat)),]

#exclusion of all-zero cells
intensity_mat_bg_corrected <- intensity_mat_mean_80[!grepl("IF2_GD2|IF3_CD45/56|DAPI", rownames(intensity_mat_mean_80)),]
non_zeros <- as.logical(colSums(intensity_mat_bg_corrected)!=0)

length(non_zeros[non_zeros==FALSE])

spe <- spe[, non_zeros]
mat <- intensity_mat_mean_80[, non_zeros]
cell_id <- colnames(spe)

#Consideration of only IMC channels and nuclear IF channels for identification of further lost cells
mat <- mat[!grepl("IF2_GD2|IF3_CD45/56", rownames(mat)),]

```
We are excluding 6121 due to zero intensity. 

In a next step, we will plot a single-cell heatmap using complex heatmap. 

```{r single-cell-heatmap, message=FALSE, fig.height=20}
library(ComplexHeatmap)
library(viridis)
library(circlize)
library(RColorBrewer)
library(paletteer)

mat <- mat

qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector_74 = unlist(mapply(brewer.pal, qual_col_pals$maxcolors, rownames(qual_col_pals)))
col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]

col_sample = setNames(sample(col_vector_433, length(unique(spe$sample))), unique(spe$sample))

set.seed(221228)
rand_cells <- sample(seq_len(ncol(mat)), 2000)

column_ha = HeatmapAnnotation(sample = colnames(mat[,rand_cells]), col=list(sample=col_sample), show_legend=F)
Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap",
        top_annotation = column_ha, 
        show_column_dend = F,
        show_column_names=F,
        col=viridis(100)
        )
```

We can already see a cluster of cells expressing DAPI but not IMC markers. We can however not exclude that those cells that have a weak Iridium staining are showing the signal from neighboring cells. 

We will now cluster the data to identify cells that swam off. 

```{r, echo=FALSE}
start_time <- Sys.time()
```

```{r phenograph-clustering}
library(Rphenoannoy)
mat <- t(mat)

set.seed(230619)

Rphenograph_out <- Rphenoannoy(mat, k = 45)

clusters <- factor(membership(Rphenograph_out[[2]]))

spe$pg_clusters_lostcells <- clusters
```

```{r, echo=FALSE}
end_time <- Sys.time()

end_time-start_time
```

RPhenograph took about 56 min. Next we will look at the single cell heatmap ordered by the identified phenograph clusters. 

```{r single-cell-heatmap-clusters, message=FALSE, fig.height=17}
library(ComplexHeatmap)
library(viridis)
library(circlize)
library(RColorBrewer)
library(paletteer)

set.seed(221228)
n <- length(unique(clusters))
col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]
col_clusters <- setNames(sample(col_vector_433, n), unique(clusters))

col_marker = colorRamp2(c(0, 2.5, 0.5), c("blue", "white", "red"))

mat <- t(mat)
rand_cells <- sample(seq_len(ncol(mat)), 2000)

column_ha = HeatmapAnnotation(cluster = clusters[rand_cells], 
                              DNA = mat["DNA1_Ir193_mean-80", rand_cells],
                              col=list(cluster=col_clusters, DNA=col_marker), 
                              show_legend=F)


Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap witquh phenograph clusters (nodil -mean)",
        top_annotation = column_ha, 
        cluster_columns=F,
        column_order= order(clusters[rand_cells], decreasing=F),
        show_column_dend = F,
        show_column_names=F,
        col=viridis(100)
        )
```

Next, we will look at the mean marker expression per cluster. 

```{r ingle-cell-heatmap-clusters-mean, message=FALSE, fig.height=17}
library(scuttle)



mean <- aggregate(t(mat), list(clusters), mean)
mean_t <- t(mean)
colnames(mean_t) <- mean[,"Group.1"]

column_ha = HeatmapAnnotation(cluster = mean[,"Group.1"], 
                              ncells=anno_barplot(as.data.frame(table(clusters))[,"Freq"], height = unit(3, "cm")),
                              col=list(cluster=col_clusters), 
                              show_legend=T)

mean_t <- t(mean[,2:ncol(mean)])
colnames(mean_t) <- mean[,"Group.1"]

Heatmap(mean_t, 
        column_title = "Cluster-means heatmap with phenograph clusters (no-dil -mean)",
        top_annotation = column_ha, 
        cluster_columns=T,
        show_column_dend = F,
        show_column_names=T,
        show_row_names=T,
        col=viridis(100)
        )


```
We can see that cluster 34 are most probably the sticky cells, which are positive for everything. Cluster 11 are most probably cells that swam off between IF and IMC and are hence negative for everything except for DAPI. We will check which sample has the highest presence of this cluster.

```{r identify-zeros-clusters, fig.width=20}

c <- "11"

data <- data.frame(sample=colnames(mat), clusters)
data$freq <- 1

data_wide <- data %>% pivot_wider(names_from = c("clusters"), values_from = "freq", values_fn = sum, values_fill = 0)
data_long <- data_wide %>% pivot_longer(cols = !sample, names_to = "cluster", values_to = "count") 

clust <- data_long[data_long$cluster==c,]
cells_per_sample <- table(spe$sample)[clust$sample]
lost_cells_perc <- clust$count/cells_per_sample
clust$percentage <- lost_cells_perc
clust <- clust[order(clust$percentage, decreasing=T), ]
max_sample <- clust$sample[1]

df <- as.data.frame(t(mat))
df$cluster <- clusters
df$sample <- colnames(mat)
rownames(df) <- cell_id


rownames(df)[df$sample==max_sample & df$cluster ==c]

write.csv(clust, file.path(path,"results/Manuscript/1_Figure/lost_cells.csv"))


```
Sample 20220316_11-1272_BM_N_ROI_001 has the highest presence of cluster 11. Let's now color its mask by the clusters. Therefore we will just copy the img and mask of that sample into the "img" and "masks" folder as reading in all images takes too long. 

```{r visualize-clusters-in-image, message=FALSE, fig.width=15}
library(cytomapper)

spe$pg_clusters <- clusters

masks <- loadImages(file.path(path, "masks_temp"), as.is = TRUE)

mcols(masks) <- DataFrame(sample_id = names(masks))


c = "11"
celltype <- spe[,spe$pg_clusters == c]

#Number of excluded cells
ncol(celltype)

plotCells(masks,
          object = celltype, 
          cell_id = "ObjectNumber", img_id = "sample_id",
          colour_by = "pg_clusters",
          colour = list(celltype = c(celltype = "red")),
          missing_colour = "white",
          display="single",
          #save_plot = list(filename = file.path(path, paste0("results/20220701_20220214_11-1272_BM_001_cluster_", c, ".png"))),
          legend=NULL)

```

By overlaying the DAPI image, the Iridium image and the binary mask, we see that cluster 11 is the cluster of lost cells (n=30285 -1.5% of all (n=2033550) cells - 648images, on average 3138 cells per image  i.e. 46 cells excluded per image). We will exclude those cells.

```{r exclude-lost-cells-in-1-px}

spe <- spe[, spe$pg_clusters!="11"]
saveRDS(spe, file.path(path, "results/Manuscript/1_Figure/spe_0px_wo-lostcells.rds"))

```
