==> code/R/20231214_Manuscript/1_Figure/04_integration_morph-features.Rmd <==
---
title: "Integration of morphological features"
output: html_document
date: '2023-05-18'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ggplot2)
library(scuttle)
library("FactoMineR")
library("factoextra")
library(bruceR)
library(SpatialExperiment)
library(cytomapper)
```


First, we will read the SPE object and define a matrix containing only morphological features area, solidity and aspect_ratio as these are the least correlated features.

```{r read-data}

path <- "/mnt/Multimodal_Imaging_Daria/PIPELINE/NEWEST_RESULTS"

spe <- readRDS(file.path(path,"results/Manuscript/1_Figure/spe_0px_wo-lostcells.rds"))

spe_all <- spe
tissue <- spe$tissue
spe_bm <- spe[,tissue=="BM"]
spe_pt <- spe[,tissue=="PT"]

morph_names <- c("area",
                 "solidity",
                 "aspect_ratio-0"
                 )

morph_mat <- as.matrix(t(colData(spe)[,morph_names]))
rownames(morph_mat) <- gsub('-0','',rownames(morph_mat))

```


## Outlier exclusion based on manual thresholding

In the distribution of morphological features, we can see that we have a lot of outliers. 
```{r histograms-separate-x-axis, message=FALSE, fig.height=3, fig.width=7}


plot_list <- list()

df <- as.data.frame(t(morph_mat))

for(i in 1: ncol(df)){

  feature <- colnames(df)[i]
  
  plot_list[[i]] <- ggplot(df, aes_string(feature)) +
  geom_density(adjust = 5, color="darkblue", fill="lightblue")
}

plot <- plot_grid(plotlist = plot_list, nrow=ceiling(ncol(df)/3), ncol=3) #change this to the number of morphological features

plot
```

We will now identify feature-individual thresholds. 

```{r quality-control-outliers, message=FALSE}

colnames(morph_mat) <- colnames(spe)

feat <- "area"

morph_mat_order <- morph_mat[feat,order(morph_mat[feat,], decreasing=F)]
outlier_cell <- names(morph_mat_order[5000])
position <- spatialCoords(spe[,outlier_cell])
print(paste("The outlier cell is in sample", outlier_cell, "at position x=", position[1], "and y=", position[2], "."))
```

We will add another feature to the SPE object for outlier annotation.

```{r annotate-outliers, message=FALSE}
sum(spe_bm$area <= 250)

spe_bm$outlier <- "no"
spe_bm[,spe_bm$area <= 250]$outlier <- "yes"

sum(spe_pt$area <= 100)

spe_pt$outlier <- "no"
spe_pt[,spe_pt$area <= 100]$outlier <- "yes"
```
We have 7923 outliers among the BM samples and 922 among the PT samples. 

We will check which samples have the most outliers. 

```{r identify-size-outlier-samples, fig.width=20}

c <- "yes"

data <- data.frame(sample=spe_bm$sample_id, cluster = spe_bm$outlier)
data$freq <- 1

data_wide <- data %>% pivot_wider(names_from = c("cluster"), values_from = "freq", values_fn = sum, values_fill = 0)
data_long <- data_wide %>% pivot_longer(cols = !sample, names_to = "cluster", values_to = "count") 

clust <- data_long[data_long$cluster==c,]
clust <- clust[order(clust$count, decreasing=T), ]
max_sample <- clust$sample[1]

print(paste("The sample with the most outlier cells is", max_sample, "."))
```

The sample with the most outlier cells is 20220806_20220314_17-4510_BM_004. We will pseudo-color the masks of outlier cells in the sample with the highest presence. 


```{r color-outlier-masks, message=FALSE}
out_path <- path

col_celltype <- setNames(c("red", "white"), c("yes", "no"))

files <- list.files(file.path(path, "masks_temp"))

for(i in 1: length(files)){
  masks <- loadImages(file.path(path, "masks_temp", files[i]), as.is = TRUE)
  
  mcols(masks) <- DataFrame(sample_id = names(masks))
  
  
  plotCells(masks,
            object = spe_bm, 
            cell_id = "ObjectNumber", img_id = "sample_id",
            colour_by = "outlier",
            colour = list(outlier = col_celltype),
            display="single",
            save_plot = list(filename = file.path(out_path, paste0(files[i],"f"))),
            legend=NULL)
}

```

Upon visualization in the images, we can confirm that the outliers are indeed outliers. Now we will exclude them. 

```{r exclude-outliers, message=FALSE}

# Remove segmentation artifacts

spe <- spe_all
outliers <- (spe$area <= 100 & spe$tissue == "PT" | spe$area <= 250 & spe$tissue == "BM" | spe$area >= 15600 & spe$tissue == "BM")
sum(outliers)
spe <- spe[,!outliers]

saveRDS(spe, file.path(path, "results/Manuscript/1_Figure/spe_0px-curated.rds"))

#Rebuild the matrix of morphological features of the remaining cells
morph_mat <- as.matrix(t(colData(spe)[,morph_names]))
rownames(morph_mat) <- gsub('-0','',rownames(morph_mat))
```

We are excluding 8847 segmentation artifacts based on morphology. 

## Normalization of morphological features


```{r read-data}

norm_minmax <- function(x){(x- min(x)) /(max(x)-min(x))}

robust_scalar<- function(x){(x- median(x)) /(quantile(x,probs = .75)-quantile(x,probs = .25))}

clipper <- function(data, lower, upper) {
  data[data > quantile(data, probs = c(upper))] <- quantile(data, probs = c(upper))
  data[data < quantile(data, probs = c(lower))] <- quantile(data, probs = c(lower))
  clipped_data <- norm_minmax(data)
  return(clipped_data)}

morph_mat_minmax <- t(apply(morph_mat, 1, norm_minmax))
morph_mat_robust <- t(apply(morph_mat, 1, robust_scalar))
morph_mat_scaled <- t(scale(t(morph_mat), center=T, scale=T))

tissue <- spe$tissue

area_upper <- 0.99999
solidity_lower <- 0.001
aspect_ratio_upper <- 0.9995 

clipped_area <- c(clipper(morph_mat[1,tissue=="BM"], 0, area_upper), clipper(morph_mat[1,tissue=="PT"], 0, area_upper))
clipped_area <- clipped_area[colnames(spe)]

morph_mat_clipped <- rbind(clipped_area, clipper(morph_mat[2,], solidity_lower, 1), clipper(morph_mat[3,], 0, aspect_ratio_upper))

colnames(morph_mat) <- spe$sample
colnames(morph_mat_clipped) <- spe$sample

rownames(morph_mat_clipped) <- c("area", "solidity", "aspect_ratio")

```

We will now look at the distribution of the morphological features. 

```{r histograms-separate-x-axis, message=FALSE, fig.height=3, fig.width=7}

plot_list <- list()

df <- as.data.frame(t(morph_mat_clipped))

for(i in 1: ncol(df)){

  feature <- colnames(df)[i]
  
  plot_list[[i]] <- ggplot(df, aes_string(feature)) +
  geom_density(adjust = 5, color="darkblue", fill="lightblue")
}

plot <- plot_grid(plotlist = plot_list, nrow=ceiling(ncol(df)/3), ncol=3) #change this to the number of morphological features

plot
```

We will now visualize the masks in a few samples colored in the value of the corresponding morphological feature to see if the transformation makes sense.

```{r visualize-clusters-in-image, message=FALSE, fig.width=15}
library(cytomapper)
library(viridis)
library(circlize)

files <- list.files(file.path(path, "masks_temp"))

for(i in 1: length(files)){

  masks <- loadImages(file.path(path, "masks_temp", files[i]), as.is = TRUE)
  
  mcols(masks) <- DataFrame(sample_id = names(masks))
  
  for (m in 1:length(morph_names)){
    
    spe$m <- morph_mat_clipped[m,]

    plotCells(masks,
              object = spe,
              cell_id = "ObjectNumber", img_id = "sample_id",
              colour_by = "m",
              save_plot = list(filename = file.path(path, "results", paste0(morph_names[m], '_', files[i], 'f'))),
              )
    
  }
}

```

Next, we will generate pseudobulk data and look at the samples with the highest or lowest mean value for each morphological feature. 

```{r pseudobulk, message=FALSE}

library(Matrix)


morph_pseudobulk <- aggregate(t(morph_mat_clipped), list(colnames(morph_mat_clipped)), mean)
rownames(morph_pseudobulk) <- morph_pseudobulk$Group.1
morph_pseudobulk <- morph_pseudobulk[,colnames(morph_pseudobulk) != "Group.1"]

write.csv(morph_pseudobulk, file.path(path, "results/Manuscript/1_Figure/morph_pseudobulk.csv"))

```

Next we will read in the single-cell data computed on the dilated masks and exclude the cells there as well. 
##Save data
```{r pseudobulk, message=FALSE}
spe_1px <- readRDS(file.path(path, "results/Manuscript/1_Figure/spe_1px_refined.rds"))
spe_1px <- spe_1px[,colnames(spe)]

saveRDS(spe_1px, file.path(path, "results/Manuscript/1_Figure/spe_1px-curated.rds"))
```
