==> code/R/20231214_Manuscript/1_Figure/05_cell_phenotyping.Rmd <==
---
title: "UMAP and histogram on matrices"
output: html_document
date: '2022-12-28'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ggplot2)
```

In this RMD file, we will perform clustering on the curated data (after excluding cells that swam off between IF and IMC and segmentation artifacts). First, we will read our SPE.

```{r read-data}

path <- "/mnt/Multimodal_Imaging_Daria/PIPELINE/NEWEST_RESULTS"
spe <- readRDS(file.path(path,"results/Manuscript/1_Figure/spe_1px-curated.rds"))

```


Then we will generate a matrix from intensity features and exclude low-quality markers such as IDO, PNMT, FN1, CXCR2, ClCasp (because it does not define one cell type, but is expressed across all) and all IF and nuclear markers. We will then use mean intensity for clustering, because the effect of lateral spillover or artifacts is diminished in this metric as compared to mean-80.

```{r generate-dataframes, message=FALSE, fig.height=7}
library(bruceR)
library(SpatialExperiment)

intensity_mat <- assay(spe, "exprs")
colnames(intensity_mat) <- colnames(spe)
intensity_mat <- intensity_mat[!grepl("mean-80", rownames(intensity_mat)),]


intensity_mat_hq <- intensity_mat[!grepl("PARP|IDO|DAPI|IF2_GD2|IF3_CD45/56|DNA1|H4K12Ac|H3K9Ac|CXCR2|PNMT|Fibro", rownames(intensity_mat)),]
intensity_mat_lq <- intensity_mat[grepl("PARP|IDO|DAPI|IF2_GD2|IF3_CD45/56|DNA1|H4K12Ac|H3K9Ac|CXCR2|PNMT|Fibro", rownames(intensity_mat)),]

#Make sure to exclude cells that are zero for all markers as PG won't work otherwise 
non_zeros <- as.logical(colSums(intensity_mat_hq)!=0)

intensity_mat_hq_wo_zeros <- intensity_mat_hq[,non_zeros]
intensity_mat_lq_wo_zeros <- intensity_mat_lq[,non_zeros]

#Apply changes to SPE object
spe <- spe[,non_zeros]
rowData(spe)$use_channel <- !grepl("mean-80|PARP|IDO|DAPI|IF2_GD2|IF3_CD45/56|DNA1|H4K12Ac|H3K9Ac|CXCR2|PNMT|Fibro", rownames(spe))

tissue <- spe$tissue
cell_id <- colnames(spe)
```

We are excluding 1997 more cells here, as they are negative for all markers used in downstream analysis. We will scale morphological features according to the determined thresholds. 
```{r generate-dataframes, message=FALSE, fig.height=10, fig.width=7}

#Normalization and scaling functions 
norm_minmax <- function(x){(x- min(x)) /(max(x)-min(x))}

robust_scalar<- function(x){(x- median(x)) /(quantile(x,probs = .75)-quantile(x,probs = .25))}

clipper <- function(data, lower, upper) {
  data[data > quantile(data, probs = c(upper))] <- quantile(data, probs = c(upper))
  data[data < quantile(data, probs = c(lower))] <- quantile(data, probs = c(lower))
  clipped_data <- norm_minmax(data)
  return(clipped_data)}


#Generation of matrix with morphological features
morph_names <- c("area",
                 "solidity",
                 "aspect_ratio-0"
                 )

morph_mat <- as.matrix(t(colData(spe)[,morph_names]))
rownames(morph_mat) <- gsub('-0','',rownames(morph_mat))

area_upper <- 0.99999
solidity_lower <- 0.001
aspect_ratio_upper <- 0.9995 

clipped_area <- c(clipper(morph_mat[1,tissue=="BM"], 0, area_upper), clipper(morph_mat[1,tissue=="PT"], 0, area_upper))
clipped_area <- clipped_area[colnames(spe)]

morph_mat_clipped <- rbind(clipped_area, clipper(morph_mat[2,], solidity_lower, 1), clipper(morph_mat[3,], 0, aspect_ratio_upper))

colnames(morph_mat_clipped) <- colnames(spe)

rownames(morph_mat_clipped) <- c("area", "solidity", "aspect_ratio")

#Concatenate intensity and morphological features or use intensity features only
mat <- intensity_mat_hq_wo_zeros

#Apply changes to SPE object
spe$area_clipped <- unname(morph_mat_clipped["area",])
spe$solidity_clipped <- unname(morph_mat_clipped["solidity",])
spe$aspect_ratio_clipped <- unname(morph_mat_clipped["aspect_ratio",])
```

In a next step, we will plot a single-cell heatmap using complex heatmap. 

```{r single-cell-heatmap, message=FALSE, fig.height=20}
library(ComplexHeatmap)
library(viridis)
library(circlize)
library(RColorBrewer)
library(paletteer)

set.seed(221228)

qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector_74 = unlist(mapply(brewer.pal, qual_col_pals$maxcolors, rownames(qual_col_pals)))
col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]

col_sample = setNames(sample(col_vector_433, length(unique(spe$sample))), unique(spe$sample))

rand_cells <- sample(seq_len(ncol(mat)), 2000)

Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap",
        show_column_dend = F,
        show_column_names=F,
        col=viridis(100)
        )
```

We will then reduce dimensionality of the data to 5 dimensions as described [here](https://umap-learn.readthedocs.io/en/latest/clustering.html).

```{r UMAP, message=FALSE}
library(scater)
library(BiocParallel)

#mat_red_clustering <- calculateUMAP(mat, ncomponents=5, n_neighbors=30, min_dist= 0, BPPARAM = MulticoreParam(RNGseed = 221228))
rownames(mat_red_clustering) <- colnames(spe)

#saveRDS(mat_red_clustering, file.path(path, "results/Manuscript/1_Figure/mat_red_clustering.rds"))

mat_red_clustering <- readRDS(file.path(path,"results/Manuscript/1_Figure/mat_red_clustering.rds"))

#Apply changes to SPE object
reducedDim(spe, "UMAP_clustering") <- mat_red_clustering
```

We will now cluster the dimensionality reduced data.

```{r, echo=FALSE}
start_time <- Sys.time()
```

```{r phenograph-clustering}
library(Rphenoannoy)

set.seed(221228)

#Rphenograph_out <- Rphenoannoy(mat_red_clustering, k = 45)

#clusters <- factor(membership(Rphenograph_out[[2]]))
#saveRDS(clusters, file.path(path, "results/Manuscript/1_Figure/pg_clusters.rds"))

clusters <- readRDS(file.path(path, "results/Manuscript/1_Figure/pg_clusters.rds"))
spe$pg_clusters <- clusters
```

```{r, echo=FALSE}
end_time <- Sys.time()

end_time-start_time
```

Next we will look at the single cell heatmap ordered by the identified phenograph clusters for 2000 randomly selected cells. 

```{r single-cell-heatmap-clusters, message=FALSE, fig.height=17}
library(ComplexHeatmap)
library(viridis)
library(circlize)
library(RColorBrewer)
library(paletteer)


set.seed(20230521)

col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]
col_clusters = setNames(sample(col_vector_433, length(unique(spe$pg_clusters))), unique(spe$pg_clusters))

col_marker = colorRamp2(c(0, 2.5, 0.5), c("blue", "white", "red"))
col_tissue = c("PT" = "deepskyblue", "BM" = "orange")

rand_cells <- sample(seq_len(ncol(mat)), 2000)


area = morph_mat_clipped["area",]
solidity = morph_mat_clipped["solidity",]
aspect_ratio = morph_mat_clipped["aspect_ratio",]

clPARP = intensity_mat_lq_wo_zeros["clPARP_Yb176_mean",]
FN = intensity_mat_lq_wo_zeros["Fibronectin_La139_mean",]

col_area = colorRamp2(c(min(area[rand_cells]),min(area[rand_cells])+(max(area[rand_cells])-min(area[rand_cells]))/2, max(area[rand_cells])), c("blue", "white", "red"))
col_solidity = colorRamp2(c(min(solidity[rand_cells]),min(solidity[rand_cells])+(max(solidity[rand_cells])-min(solidity[rand_cells]))/2, max(solidity[rand_cells])), c("blue", "white", "red"))
col_aspect_ratio = colorRamp2(c(min(aspect_ratio[rand_cells]),min(aspect_ratio[rand_cells])+(max(aspect_ratio[rand_cells])-min(aspect_ratio[rand_cells]))/2, max(aspect_ratio[rand_cells])), c("blue", "white", "red"))

col_clPARP = colorRamp2(c(min(clPARP[rand_cells]),min(clPARP[rand_cells])+(max(clPARP[rand_cells])-min(clPARP[rand_cells]))/2, max(clPARP[rand_cells])), c("blue", "white", "red"))
col_FN = colorRamp2(c(min(FN[rand_cells]),min(FN[rand_cells])+(max(FN[rand_cells])-min(FN[rand_cells]))/2, max(FN[rand_cells])), c("blue", "white", "red"))

column_ha = HeatmapAnnotation(cluster = clusters[rand_cells],
                              tissue = tissue[rand_cells],
                              area = area[rand_cells],
                              solidity = solidity[rand_cells],
                              aspect_ratio = aspect_ratio[rand_cells],
                              clPARP = clPARP[rand_cells],
                              FN = FN[rand_cells],
                              col=list(cluster=col_clusters, 
                                       tissue=col_tissue,
                                       area = col_area,
                                       solidity = col_solidity,
                                       aspect_ratio = col_aspect_ratio,
                                       clPARP = col_clPARP,
                                       FN = col_FN
                                       ), 
                              show_legend=T)


Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap with phenograph clusters (1-px mean)",
        top_annotation = column_ha, 
        cluster_columns=F,
        column_order= order(clusters[rand_cells], decreasing=F),
        show_column_dend = T,
        show_column_names=F,
        col=viridis(100)
        )
```
Next, we will plot this heatmap for every cluster separately.
```{r single-cell-heatmap-clusters, message=FALSE, fig.height=17}

set.seed(20230521)

mat_hq <- mat
mat_lq <- intensity_mat_lq_wo_zeros
mat_morph <- morph_mat_clipped
clusters <- spe$pg_clusters

for(i in 1: length(unique(clusters))){

  c <- unique(clusters)[i]

  mat_c <- mat_hq[,clusters==c]
  ct <- clusters[clusters==c]
  tis <- spe$tissue[clusters==c]
  
  col_marker = colorRamp2(c(0, 2.5, 0.5), c("blue", "white", "red"))
  col_tissue = c("PT" = "deepskyblue", "BM" = "orange")
  
  if (length(ct) < 8000){len=length(ct)}else{len=8000}
  rand_cells <- sample(seq_len(ncol(mat_c)), len)
  
  
  area_c = mat_morph["area",clusters==c]
  solidity_c = mat_morph["solidity",clusters==c]
  aspect_ratio_c = mat_morph["aspect_ratio",clusters==c]
  
  clPARP_c = mat_lq["clPARP_Yb176_mean",clusters==c]
  FN_c = mat_lq["Fibronectin_La139_mean",clusters==c]
  
  col_area_c = colorRamp2(c(min(area_c[rand_cells]),min(area_c[rand_cells])+(max(area_c[rand_cells])-min(area_c[rand_cells]))/2, max(area_c[rand_cells])), c("blue", "white", "red"))
  col_solidity_c = colorRamp2(c(min(solidity_c[rand_cells]),min(solidity_c[rand_cells])+(max(solidity_c[rand_cells])-min(solidity_c[rand_cells]))/2, max(solidity_c[rand_cells])), c("blue", "white", "red"))
  col_aspect_ratio_c = colorRamp2(c(min(aspect_ratio_c[rand_cells]),min(aspect_ratio_c[rand_cells])+(max(aspect_ratio_c[rand_cells])-min(aspect_ratio_c[rand_cells]))/2, max(aspect_ratio_c[rand_cells])), c("blue", "white", "red"))
  
  col_clPARP_c = colorRamp2(c(min(clPARP_c[rand_cells]),min(clPARP_c[rand_cells])+(max(clPARP_c[rand_cells])-min(clPARP_c[rand_cells]))/2, max(clPARP_c[rand_cells])), c("blue", "white", "red"))
  col_FN_c = colorRamp2(c(min(FN_c[rand_cells]),min(FN_c[rand_cells])+(max(FN_c[rand_cells])-min(FN_c[rand_cells]))/2, max(FN_c[rand_cells])), c("blue", "white", "red"))
  
  column_ha = HeatmapAnnotation(clusters = ct[rand_cells],
                                tissue = tis[rand_cells],
                                area = area_c[rand_cells],
                                solidity = solidity_c[rand_cells],
                                aspect_ratio = aspect_ratio_c[rand_cells],
                                clPARP = clPARP_c[rand_cells],
                                FN = FN_c[rand_cells],
                                col=list(clusters=col_clusters, 
                                         tissue=col_tissue,
                                         area = col_area_c,
                                         solidity = col_solidity_c,
                                         aspect_ratio = col_aspect_ratio_c,
                                         clPARP = col_clPARP_c,
                                         FN = col_FN_c
                                         ), 
                                show_legend=T)
  
  png(file=file.path(path, paste0("results/Manuscript/1_Figure/", c , ".tiff")),width=18,height=8,units="in", res=1200)
  
  ht<- Heatmap(mat_c[,rand_cells], 
          column_title = "Single-cell heatmap with 8000 randomly selected cells",
          top_annotation = column_ha, 
          cluster_columns=F,
          column_order= order(tis[rand_cells], decreasing=F),
          show_column_dend = T,
          show_column_names=F,
          col=viridis(100)
          )
  draw(ht)
  
  dev.off()
}

```

Next, we will plot this heatmap for every cluster and tissue separately.
```{r single-cell-heatmap-clusters, message=FALSE, fig.height=17}

set.seed(20230521)

t = "BM"

mat_hq <- mat
mat_lq <- intensity_mat_lq_wo_zeros
mat_morph <- morph_mat_clipped
clusters <- spe$pg_clusters
tissue <- spe$tissue

for(i in 1: length(unique(clusters))){

  c <- unique(clusters)[i]

  mat_c <- mat_hq[,clusters==c & tissue ==t]
  ct <- clusters[clusters==c & tissue ==t]
  tis <- spe$tissue[clusters==c & tissue ==t]
  
  col_marker = colorRamp2(c(0, 2.5, 0.5), c("blue", "white", "red"))
  col_tissue = c("PT" = "deepskyblue", "BM" = "orange")
  
  if (length(ct) < 8000){len=length(ct)}else{len=8000}
  if (length(ct) <2){next}
  rand_cells <- sample(seq_len(ncol(mat_c)), len)
  
  area_c = mat_morph["area",clusters==c & tissue ==t]
  solidity_c = mat_morph["solidity",clusters==c & tissue ==t]
  aspect_ratio_c = mat_morph["aspect_ratio",clusters==c & tissue ==t]
  
  clPARP_c = mat_lq["clPARP_Yb176_mean",clusters==c & tissue ==t]
  FN_c = mat_lq["Fibronectin_La139_mean",clusters==c & tissue ==t]
  
  col_area_c = colorRamp2(c(min(area_c[rand_cells]),min(area_c[rand_cells])+(max(area_c[rand_cells])-min(area_c[rand_cells]))/2, max(area_c[rand_cells])), c("blue", "white", "red"))
  col_solidity_c = colorRamp2(c(min(solidity_c[rand_cells]),min(solidity_c[rand_cells])+(max(solidity_c[rand_cells])-min(solidity_c[rand_cells]))/2, max(solidity_c[rand_cells])), c("blue", "white", "red"))
  col_aspect_ratio_c = colorRamp2(c(min(aspect_ratio_c[rand_cells]),min(aspect_ratio_c[rand_cells])+(max(aspect_ratio_c[rand_cells])-min(aspect_ratio_c[rand_cells]))/2, max(aspect_ratio_c[rand_cells])), c("blue", "white", "red"))
  
  column_ha = HeatmapAnnotation(clusters = ct[rand_cells],
                                tissue = tis[rand_cells],
                                area = area_c[rand_cells],
                                solidity = solidity_c[rand_cells],
                                aspect_ratio = aspect_ratio_c[rand_cells],
                                col=list(clusters=col_clusters, 
                                         tissue=col_tissue,
                                         area = col_area_c,
                                         solidity = col_solidity_c,
                                         aspect_ratio = col_aspect_ratio_c
                                         ), 
                                show_legend=T)
  
  png(file=file.path(path, paste0("results/Manuscript/1_Figure/", t, "/", t, "_" , c , ".tiff")),width=18,height=8,units="in", res=1200)
  
  ht<- Heatmap(mat_c[,rand_cells], 
          column_title = "Single-cell heatmap with 8000 randomly selected cells",
          top_annotation = column_ha, 
          cluster_columns=F,
          column_order= order(tis[rand_cells], decreasing=F),
          show_column_dend = T,
          show_column_names=F,
          col=viridis(100)
          )
  draw(ht)
  
  dev.off()
}
```


Next we will look at the single cell heatmap ordered by the identified phenograph clusters and sample the same number of cells from BM and PT samples. 

```{r single-cell-heatmap-clusters, message=FALSE, fig.height=10}
library(ComplexHeatmap)
library(viridis)
library(circlize)
library(RColorBrewer)
library(paletteer)

set.seed(20230618) 

col_marker = colorRamp2(c(0, 2.5, 0.5), c("blue", "white", "red"))
col_tissue = c("PT" = "deepskyblue", "BM" = "orange")

#Get same number of PT and BM cells
pt_index <- which(tissue=="PT")
bm_index <- which(tissue=="BM")
rand_cells_pt <- sample(seq_len(length(pt_index)), 4000)
rand_cells_bm <- sample(seq_len(length(bm_index)), 4000)
pt_index_rand <- pt_index[rand_cells_pt]
bm_index_rand <- bm_index[rand_cells_bm]
rand_cells <- c(pt_index_rand, bm_index_rand)

col_area = colorRamp2(c(min(area[rand_cells]),min(area[rand_cells])+(max(area[rand_cells])-min(area[rand_cells]))/2, max(area[rand_cells])), c("blue", "white", "red"))
col_solidity = colorRamp2(c(min(solidity[rand_cells]),min(solidity[rand_cells])+(max(solidity[rand_cells])-min(solidity[rand_cells]))/2, max(solidity[rand_cells])), c("blue", "white", "red"))
col_aspect_ratio = colorRamp2(c(min(aspect_ratio[rand_cells]),min(aspect_ratio[rand_cells])+(max(aspect_ratio[rand_cells])-min(aspect_ratio[rand_cells]))/2, max(aspect_ratio[rand_cells])), c("blue", "white", "red"))

col_clPARP = colorRamp2(c(min(clPARP[rand_cells]),min(clPARP[rand_cells])+(max(clPARP[rand_cells])-min(clPARP[rand_cells]))/2, max(clPARP[rand_cells])), c("blue", "white", "red"))

column_ha = HeatmapAnnotation(cluster = clusters[rand_cells],
                              tissue = tissue[rand_cells],
                              area = area[rand_cells],
                              solidity = solidity[rand_cells],
                              aspect_ratio = aspect_ratio[rand_cells],
                              clPARP = clPARP[rand_cells],
                              col=list(cluster=col_clusters, 
                                       tissue=col_tissue,
                                       area = col_area,
                                       solidity = col_solidity,
                                       aspect_ratio = col_aspect_ratio,
                                       clPARP = col_clPARP
                                       ), 
                              show_legend=T)


Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap with phenograph clusters (1-px mean)",
        top_annotation = column_ha, 
        cluster_columns=F,
        column_order= order(clusters[rand_cells], decreasing=F),
        show_column_dend = T,
        show_column_names=F,
        col=viridis(100)
        )
```
Next we will look at the single cell heatmap ordered by the identified phenograph clusters and sample the same number of cells from BM and PT samples per cluster. 

```{r single-cell-heatmap-clusters, message=FALSE, fig.height=10}

set.seed(20230517)

n <- length(unique(clusters))

count <- 45
rand_cells <- c()

#Get same number of PT and BM cells per cluster
for (i in 1: length(unique(clusters))){
  pt_index <- which(tissue=="PT" & clusters == unique(clusters)[i])
  bm_index <- which(tissue=="BM" & clusters == unique(clusters)[i])
  if (length(pt_index) < count){
    pt_diff <- count-length(pt_index)
    print(paste("PT_diff: ", pt_diff, " for cluster ", unique(clusters)[i]))
    rand_index_pt <- sample(pt_index, length(pt_index))
    rand_index_bm <- sample(bm_index, (count + pt_diff))
  } else if (length(bm_index) < count){
    bm_diff <- count-length(bm_index)
    rand_index_pt <- sample(pt_index, (count + bm_diff))
    rand_index_bm <- sample(bm_index, length(bm_index))
  } else {
      rand_index_pt <- sample(pt_index, count)
      rand_index_bm <- sample(bm_index, count)
  }

  rand_cells <- c(rand_cells, rand_index_pt, rand_index_bm)
}

col_area = colorRamp2(c(min(area[rand_cells]),min(area[rand_cells])+(max(area[rand_cells])-min(area[rand_cells]))/2, max(area[rand_cells])), c("blue", "white", "red"))
col_solidity = colorRamp2(c(min(solidity[rand_cells]),min(solidity[rand_cells])+(max(solidity[rand_cells])-min(solidity[rand_cells]))/2, max(solidity[rand_cells])), c("blue", "white", "red"))
col_aspect_ratio = colorRamp2(c(min(aspect_ratio[rand_cells]),min(aspect_ratio[rand_cells])+(max(aspect_ratio[rand_cells])-min(aspect_ratio[rand_cells]))/2, max(aspect_ratio[rand_cells])), c("blue", "white", "red"))

col_marker = colorRamp2(c(0, 2.5, 0.5), c("blue", "white", "red"))
col_tissue = c("PT" = "deepskyblue", "BM" = "orange")

#mat <- t(mat)
                                       
                                       
column_ha = HeatmapAnnotation(cluster = clusters[rand_cells],
                              tissue = tissue[rand_cells],
                              area = area[rand_cells],
                              solidity = solidity[rand_cells],
                              aspect_ratio = aspect_ratio[rand_cells],
                              col=list(cluster=col_clusters, 
                                       tissue=col_tissue,
                                       area = col_area,
                                       solidity = col_solidity,
                                       aspect_ratio = col_aspect_ratio
                                       ), 
                              show_legend=T)


Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap with phenograph clusters (1-px mean)",
        top_annotation = column_ha, 
        cluster_columns=F,
        column_order= order(clusters[rand_cells], decreasing=F),
        show_column_dend = T,
        show_column_names=F,
        col=viridis(100)
        )
```

Next we will look at the single cell heatmap ordered by the identified phenograph clusters and sample the same number of cells from each cluster

```{r single-cell-heatmap-clusters, message=FALSE, fig.height=10}

set.seed(20230517)

n <- length(unique(clusters))

count <- 90
rand_cells <- c()

#Get same number of PT and BM cells per cluster
for (i in 1: length(unique(clusters))){
  idx <- which(clusters == unique(clusters)[i])
  rand_idx <- sample(idx, count)
  rand_cells <- c(rand_cells, rand_idx)
}

col_area = colorRamp2(c(min(area[rand_cells]),min(area[rand_cells])+(max(area[rand_cells])-min(area[rand_cells]))/2, max(area[rand_cells])), c("blue", "white", "red"))
col_solidity = colorRamp2(c(min(solidity[rand_cells]),min(solidity[rand_cells])+(max(solidity[rand_cells])-min(solidity[rand_cells]))/2, max(solidity[rand_cells])), c("blue", "white", "red"))
col_aspect_ratio = colorRamp2(c(min(aspect_ratio[rand_cells]),min(aspect_ratio[rand_cells])+(max(aspect_ratio[rand_cells])-min(aspect_ratio[rand_cells]))/2, max(aspect_ratio[rand_cells])), c("blue", "white", "red"))

col_marker = colorRamp2(c(0, 2.5, 0.5), c("blue", "white", "red"))
col_tissue = c("PT" = "deepskyblue", "BM" = "orange")

#mat <- t(mat)

column_ha = HeatmapAnnotation(cluster = clusters[rand_cells],
                              tissue = tissue[rand_cells],
                              area = area[rand_cells],
                              solidity = solidity[rand_cells],
                              aspect_ratio = aspect_ratio[rand_cells],
                              col=list(cluster=col_clusters, 
                                       tissue=col_tissue,
                                       area = col_area,
                                       solidity = col_solidity,
                                       aspect_ratio = col_aspect_ratio
                                       ), 
                              show_legend=T)


Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap with phenograph clusters (1-px mean)",
        top_annotation = column_ha, 
        cluster_columns=F,
        column_order= order(clusters[rand_cells], decreasing=F),
        show_column_dend = T,
        show_column_names=F,
        col=viridis(100)
        )
```

Next, we will look at the mean marker expression per cluster. The next two code blocks can be also run for tissue types (PT/BM) separately. 

```{r ingle-cell-heatmap-clusters-mean, message=FALSE, fig.height=10}
library(scuttle)

set.seed(20230135)

#calculate mean heatmap over all cells
matrix_hq <- intensity_mat_hq_wo_zeros
matrix_lq <- intensity_mat_lq_wo_zeros
morph_matrix <- morph_mat_clipped
clust <- droplevels(spe$pg_clusters)

#calculate heatmap over tissue-derived cells
# tissue <- "BM"
# matrix_hq <- intensity_mat_hq_wo_zeros[,spe$tissue==tissue]
# matrix_lq <- intensity_mat_lq_wo_zeros[,spe$tissue==tissue]
# morph_matrix <- morph_mat_clipped[,spe$tissue==tissue]
# clust <- clusters[spe$tissue==tissue]

mean <- aggregate(t(matrix_hq), list(clust), mean)

mean_t <- t(mean[,2:ncol(mean)])
colnames(mean_t) <- mean[,"Group.1"]

mean_morph <- aggregate(t(morph_matrix), list(clust), mean)

area_mean = mean_morph[,"area"]
solidity_mean = mean_morph[,"solidity"]
aspect_ratio_mean = mean_morph[,"aspect_ratio"]

mean_lq <- aggregate(t(matrix_lq), list(clust), mean)

clPARP_mean = mean_lq[,"clPARP_Yb176_mean"]

col_mean_area = colorRamp2(c(min(area_mean),min(area_mean)+(max(area_mean)-min(area_mean))/2, max(area_mean)), c("blue", "white", "red"))
col_mean_solidity = colorRamp2(c(min(solidity_mean),min(solidity_mean)+(max(solidity_mean)-min(solidity_mean))/2, max(solidity_mean)), c("blue", "white", "red"))
col_mean_aspect_ratio = colorRamp2(c(min(aspect_ratio_mean),min(aspect_ratio_mean)+(max(aspect_ratio_mean)-min(aspect_ratio_mean))/2, max(aspect_ratio_mean)), c("blue", "white", "red"))
col_mean_clPARP = colorRamp2(c(min(clPARP_mean),min(clPARP_mean)+(max(clPARP_mean)-min(clPARP_mean))/2, max(clPARP_mean)), c("blue", "white", "red"))

column_ha = HeatmapAnnotation(cluster = mean[,"Group.1"], 
                              area=area_mean,
                              solidity = solidity_mean,
                              aspect_ratio = aspect_ratio_mean,
                              clPARP = clPARP_mean,
                              ncells=anno_barplot(as.data.frame(table(clust))[,"Freq"], height = unit(3, "cm")),
                              col=list(cluster=col_clusters, 
                                       area=col_mean_area,
                                       solidity=col_mean_solidity,
                                       aspect_ratio=col_mean_aspect_ratio,
                                       clPARP = col_mean_clPARP
                                       ), 
                              show_legend=T)

Heatmap(mean_t, 
        column_title = "Cluster-means heatmap with phenograph clusters (1px -mean)-scaled",
        top_annotation = column_ha, 
        cluster_columns=T,
        show_column_dend = T,
        show_column_names=T,
        show_row_names=T,
        col=viridis(100)
        )


```
We can see that cluster 53 are the sticky cells, which are positive for everything. We will exclude this cluster and re-generate the heatmap.

```{r ingle-cell-heatmap-clusters-mean, message=FALSE, fig.height=12}
library(scuttle)

set.seed(20230139)

sticky_cluster <- "53"

# Feature-wise scaling
mean_t_wo_sticky <- mean_t[,levels(clusters)!=sticky_cluster]
mean_t_wo_sticky <- t(apply(mean_t_wo_sticky, 1, RESCALE, to=0:1))
clusters_wo_sticky <- droplevels(clust[clust!=sticky_cluster])

area_mean_wo_sticky <- area_mean[levels(clusters)!=sticky_cluster]
solidity_mean_wo_sticky <- solidity_mean[levels(clusters)!=sticky_cluster]
aspect_ratio_mean_wo_sticky <- aspect_ratio_mean[levels(clusters)!=sticky_cluster]
clPARP_mean_wo_sticky <- clPARP_mean[levels(clusters)!=sticky_cluster]


col_mean_area_wo_sticky = colorRamp2(c(min(area_mean_wo_sticky),min(area_mean_wo_sticky)+(max(area_mean_wo_sticky)-min(area_mean_wo_sticky))/2, max(area_mean_wo_sticky)), c("blue", "white", "red"))
col_mean_solidity_wo_sticky = colorRamp2(c(min(solidity_mean_wo_sticky),min(solidity_mean_wo_sticky)+(max(solidity_mean_wo_sticky)-min(solidity_mean_wo_sticky))/2, max(solidity_mean_wo_sticky)), c("blue", "white", "red"))
col_mean_aspect_ratio_wo_sticky = colorRamp2(c(min(aspect_ratio_mean_wo_sticky),min(aspect_ratio_mean_wo_sticky)+(max(aspect_ratio_mean_wo_sticky)-min(aspect_ratio_mean_wo_sticky))/2, max(aspect_ratio_mean_wo_sticky)), c("blue", "white", "red"))
col_mean_clPARP_wo_sticky = colorRamp2(c(min(clPARP_mean_wo_sticky),min(clPARP_mean_wo_sticky)+(max(clPARP_mean_wo_sticky)-min(clPARP_mean_wo_sticky))/2, max(clPARP_mean_wo_sticky)), c("blue", "white", "red"))

col_clusters_wo_sticky <- col_clusters[names(col_clusters)!=sticky_cluster]

column_ha = HeatmapAnnotation(cluster = droplevels(mean[levels(clusters)!=sticky_cluster,"Group.1"]),
                              area=area_mean_wo_sticky,
                              solidity=solidity_mean_wo_sticky,
                              aspect_ratio=aspect_ratio_mean_wo_sticky,
                              clPARP=clPARP_mean_wo_sticky,
                              ncells=anno_barplot(as.data.frame(table(clusters_wo_sticky))[,"Freq"], height = unit(3, "cm")),
                              col=list(cluster=col_clusters_wo_sticky,
                                       area=col_mean_area_wo_sticky,
                                       solidity=col_mean_solidity_wo_sticky,
                                       aspect_ratio=col_mean_aspect_ratio_wo_sticky,
                                       clPARP=col_mean_clPARP_wo_sticky
                                       ), 
                              show_legend=T)

Heatmap(mean_t_wo_sticky, 
        column_title = "Cluster-means heatmap with phenograph clusters (1px -mean)-scaled",
        top_annotation = column_ha, 
        row_labels=sapply(strsplit(rownames(mean_t_wo_sticky), "_"), "[[", 1),
        cluster_columns=T,
        cluster_rows=T,
        show_column_dend = T,
        show_column_names=T,
        show_row_names=T,
        col=viridis(100)
        )


```

Next we will visualize the clusters in a bubble plot. 
```{r dot-plot-complex-heatmap, message=FALSE, fig.height=14, fig.width=16}
library(tidyr)
library(tidyHeatmap)
library(bruceR)

#Mean expression per cluster

exp_mat_ <- aggregate(t(mat), list(spe$pg_clusters), mean)
exp_mat <- exp_mat_[,2:ncol(exp_mat_)]
exp_mat <- apply(exp_mat, 2, RESCALE, to=0:1)
rownames(exp_mat) <- exp_mat_[,"Group.1"]
exp_mat <- t(exp_mat)
head(exp_mat)

#Percentage of cells per cluster expressing a marker (>0)
tmp <- mat
tmp[tmp > unname(quantile(exp_mat[exp_mat>0], c(0.2)))] <- 1
tmp[tmp <= unname(quantile(exp_mat[exp_mat>0], c(0.2)))] <- 0
tmp <- aggregate(t(tmp), list(spe$pg_clusters), sum)
tmp2 <- tmp[,2:ncol(tmp)]
rownames(tmp2) <- tmp[,"Group.1"]
percent_mat <- tmp2 / unname(table(droplevels(spe$pg_clusters)))
percent_mat <- t(percent_mat)
percent_mat[percent_mat < 0.1] <- 0.1 #setting min values higher to make points in heatmap visible
head(percent_mat)

#Annotation bars
bm_count <- table(spe[,spe$tissue=="BM"]$pg_clusters)
bm_count[setdiff(colnames(exp_mat), names(bm_count))] <- 0
bm_count <- bm_count[colnames(exp_mat)]
pt_count <- table(spe[,spe$tissue=="PT"]$pg_clusters)
pt_count[setdiff(colnames(exp_mat), names(pt_count))] <- 0
pt_count <- pt_count[colnames(exp_mat)]

mean_morph <- aggregate(t(morph_mat_clipped), list(spe$pg_clusters), mean)
area_mean = mean_morph[,"area"]
solidity_mean = mean_morph[,"solidity"]
col_mean_area = colorRamp2(c(min(area_mean),min(area_mean)+(max(area_mean)-min(area_mean))/2, max(area_mean)), c("blue", "white", "red"))
col_mean_solidity = colorRamp2(c(min(solidity_mean),min(solidity_mean)+(max(solidity_mean)-min(solidity_mean))/2, max(solidity_mean)), c("blue", "white", "red"))

column_ha = HeatmapAnnotation(#pg_clusters = colnames(exp_mat),
                              roundness = solidity_mean,
                              #area = area_mean,
                              #ncells_overall=anno_barplot(as.data.frame(table(spe$pg_clusters))[,"Freq"], add_numbers=T, height = unit(2.2, "cm")),
                              ncells_PT=anno_barplot(unname(pt_count), add_numbers=F, height = unit(0.7, "cm"), gp = gpar(fill = "brown")),
                              #ncells_bm=anno_barplot(unname(bm_count), add_numbers=F, height = unit(0.7, "cm"), gp = gpar(fill = "aquamarine4")),
                              #ncells = anno_barplot(matrix(nc=2, c(unname(pt_count), unname(bm_count))), beside=T, attach=T, height = unit(2.2, "cm")),
                              col=list(#pg_clusters=metadata(spe)$color_vectors$col_clusters
                                       roundness = col_mean_solidity
                                       #area = col_mean_area
                                       ), 
                              #annotation_legend_param = list(pg_clusters = list(
                       	      #at = colnames(exp_mat)[order(mean_metaclust$x)])),
                              show_legend=T)

#Function to set dot size based on percentage expressed
cell_fun = function(j, i, x, y, w, h, fill){
          grid.rect(x = x, y = y, width = w, height = h, 
                    gp = gpar(col = NA, fill = NA))
          grid.circle(x=x,y=y,r= percent_mat[i, j]/2 * min(unit.c(w, h)),
                      gp = gpar(fill = col_fun(exp_mat[i, j]), col = NA))}

#Function to set dot color based on mean expression
col_fun = circlize::colorRamp2(c(min(exp_mat),(max(exp_mat)-min(exp_mat))/2, max(exp_mat)), viridis(100)[c(1,50,100)])

#Function to set legend size according to percentage expressed
layer_fun = function(j, i, x, y, w, h, fill){
          grid.rect(x = x, y = y, width = w, height = h, 
                    gp = gpar(col = NA, fill = NA))
          grid.circle(x=x,y=y,r= pindex(percent_mat, i, j)/1 * unit(1.5, "mm"),
                      gp = gpar(fill = col_fun(pindex(exp_mat, i, j)), col = NA))}

#Generate legend for percentage expressed
lgd_list = list(
    Legend( labels = c(0,0.25,0.5,0.75,1), title = "percentage expressed",
            graphics = list(
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0.25 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0.5 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0.75 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 1 * unit(2, "mm"),
                                               gp = gpar(fill = "black")))
            ))

#Heatmap
hp <- Heatmap(exp_mat,
        heatmap_legend_param=list(title="expression"),
        column_title = "Dotplot - normalized mean expression and %expressed", 
        top_annotation = column_ha, 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        clustering_distance_columns="pearson", #use euclidean, kendall or pearson
        clustering_method_columns="average", #use centroid, average or single
        layer_fun = layer_fun,
        row_labels=sapply(strsplit(rownames(exp_mat), "_"), "[[", 1),
        cell_fun = cell_fun,
        show_row_dend = F,
        column_dend_height = unit(1.5, "cm"),
        row_names_gp = gpar(fontsize = 10),
        column_names_gp = gpar(fontsize = 10),
        column_names_rot = 45,
        border = "black")

d1 <- draw( hp, annotation_legend_list = lgd_list)

setEPS()
postscript(file.path(path, "results/Manuscript/2_Ext_Figure/R/bubbleplot_scaled_PT.eps"), height=7.2, width=18)
d1
dev.off()
```

Then, we will perform dimensionality reduction by UMAP for visualization.

```{r UMAP, message=FALSE}

#mat_red <- calculateUMAP(mat, n_neighbors=30, BPPARAM = MulticoreParam(RNGseed = 221228))

rownames(mat_red) <- colnames(spe)

#saveRDS(mat_red, file.path(path, "results/Manuscript/1_Figure/mat_red.rds"))

mat_red <- readRDS(file.path(path,"results/Manuscript/1_Figure/mat_red.rds"))

#Apply changes to SPE object
reducedDim(spe, "UMAP") <- mat_red
```

We will first plot the clusters in the UMAP.

```{r cell type umap, fig.width=15, message=FALSE}
library(dittoSeq)
library(scater)
library(patchwork)
library(cowplot)
library(viridis)


## UMAP colored by cell type and expression - dittoDimPlot
p1 <- dittoDimPlot(spe, var = "pg_clusters", 
             reduction.use = "UMAP", size = 1,
             do.label = TRUE,
             legend.show=F) +
  scale_color_manual(values = col_clusters) + 
  theme(legend.title = element_blank()) +
  ggtitle("PG clusters on UMAP")

p1
```

Next we will visualize the tissue type in the UMAP space.

```{r visualizing-umap-by-phenograph-cluster, message=F, fig.width=10}
tissue <- spe$tissue
mat_red_df <- as.data.frame(mat_red)
mat_red_df["tissue"] <- tissue
mat_red_df_rand <- mat_red_df[sample(1:nrow(mat_red_df)),]
  
plot <- ggplot(mat_red_df_rand, aes(x=V1, y=V2, colour=tissue)) +
    geom_point(alpha=0.5, show.legend=T) + 
    ggtitle("Tissue type") +
    scale_colour_manual(values = col_tissue) +
    theme(plot.title = element_text(size = 15), 
          axis.title = element_text(size=15), 
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(), 
          panel.background = element_blank(), 
          axis.line = element_line(colour = "black"),
          legend.text = element_text(size=15))

plot
```

Next we will visualize the clusters in the UMAP space.

```{r visualizing-umap-by-phenograph-cluster, message=F, fig.width=10}

mat_red_df <- as.data.frame(mat_red)
mat_red_df["clusters"] <- clusters
mat_red_df_rand <- mat_red_df[sample(1:nrow(mat_red_df)),]
  
plot <- ggplot(mat_red_df_rand, aes(x=V1, y=V2, colour=clusters)) +
    geom_point(alpha=0.5, show.legend=T) + 
    ggtitle("Phenograph cluster") +
    scale_colour_manual(values = col_clusters) +
    theme(plot.title = element_text(size = 15), 
          axis.title = element_text(size=15), 
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(), 
          panel.background = element_blank(), 
          axis.line = element_line(colour = "black"),
          legend.text = element_text(size=15))

plot
```

Now, we will plot all samples in the dim-red space. 

```{r visualizing-umap-by-sample, message=F, fig.width=10}

library(RColorBrewer)

rownames(mat_red) <- spe$sample
mat_red_df <- as.data.frame(mat_red)
mat_red_df["sample"] <- rownames(mat_red)
mat_red_df_rand <- mat_red_df[sample(1:nrow(mat_red_df)),]
  
plot <- ggplot(mat_red_df_rand, aes(x=V1, y=V2, colour=sample)) +
    geom_point(size=0.05, alpha=0.5, show.legend=F) + 
    ggtitle("Sample") +
    scale_colour_manual(values = col_sample) +
    theme(plot.title = element_text(size = 5), 
          axis.title = element_text(size=5), 
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(), 
          panel.background = element_blank(), 
          axis.line = element_line(colour = "black"),
          legend.text = element_text(size=5))

plot
```

Now, we will plot the marker separately in the dim-red space. 

```{r visualizing-umap-by-marker, message=F, fig.width=20}
mat_tmp <- rbind(mat, intensity_mat_lq_wo_zeros, morph_mat_clipped)
#mat_tmp <- morph_mat_clipped[c("area", "solidity"),]
#mat_tmp <- mat

for (i in 1: nrow(mat_tmp)){
  marker <- rownames(mat_tmp)[i]
  mat_red_df <- as.data.frame(mat_red)
  mat_red_df["marker"] <- mat_tmp[i,]
  mat_red_df_rand <- mat_red_df[sample(1:nrow(mat_red_df)),]
  #mat_red_df_rand <- mat_red_df[order(mat_red_df["marker"], decreasing=F),]
    
  plot <- ggplot(mat_red_df_rand, aes(x=V1, y=V2, colour=marker)) +
      geom_point(size=0.05, alpha=0.2, show.legend=T) + 
      ggtitle(marker) +
      scale_colour_viridis_c() +
      theme(plot.title = element_text(size = 5), 
            axis.title = element_text(size=5), 
            panel.grid.major = element_blank(), 
            panel.grid.minor = element_blank(), 
            panel.background = element_blank(), 
            axis.line = element_line(colour = "black"),
            legend.text = element_text(size=5))
  
  ggsave(filename=file.path(path, paste0("results/Manuscript/1_Figure/", marker , ".tiff")), width = 10, height = 5) #units = "in", device='tiff', dpi=500)
} 
```

Next, we will plot the tissue type separately in the dim-red space.

```{r visualizing-umap-by-tissuetype}

var <- "tissue"

rownames(mat_red) <- spe$sample_id
meta <- colData(spe)[match(rownames(mat_red), spe$sample_id),var]

samples <- unique(meta)

bin_col <- setNames(c(adjustcolor("gray90",alpha.f=1), "palegreen4"), c(FALSE,TRUE))

plot_list <- list()

for(i in 1: length(samples)){

  sample_binary <- meta == samples[i] 
  
  background <- as.data.frame(mat_red[!sample_binary,])
  foreground <- as.data.frame(mat_red[sample_binary,])
  plot_list[[i]] <- ggplot(background, aes(x=V1, y=V2) ) +
    geom_point(color="gray90", size=0.02) +
    geom_point(data=foreground, color="palegreen4", show.legend=F, size=0.02, alpha=0.25) +
    ggtitle(samples[i]) +
    theme(plot.title = element_text(size = 5), 
          axis.title = element_text(size=5), 
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(), 
          panel.background = element_blank(), 
          axis.line = element_line(colour = "black"))
}


plot_list[[1]]
plot_list[[2]]
```

Now, we will plot the clusters separately in the dim-red space. 

```{r visualizing-umap-by-marker, message=F, fig.width=20}

mat_red_df <- as.data.frame(mat_red)
mat_red_df["tissue"] <- spe$tissue

samples <- unique(clusters)

for(i in 1: length(samples)){

  sample_binary <- clusters == samples[i] 
  
  background <- as.data.frame(mat_red_df[!sample_binary,])
  foreground <- as.data.frame(mat_red_df[sample_binary,])
  plot <- ggplot(background, aes(x=V1, y=V2) ) +
    geom_point(color="gray90", size=0.02) +
    geom_point(data=foreground, aes(x=V1, y=V2, colour=tissue), show.legend=F, size=0.02, alpha=0.25) +
    scale_colour_manual(values = col_tissue) +
    ggtitle(samples[i]) +
    theme(plot.title = element_text(size = 5), 
          axis.title = element_text(size=5), 
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(), 
          panel.background = element_blank(), 
          axis.line = element_line(colour = "black"))
  
  ggsave(filename=file.path(path, paste0("results/Manuscript/1_Figure/", samples[i] , ".tiff")), width = 10, height = 5, units = "in", device='tiff', dpi=500)
}

```


Finally, we will visualize the data in the dimensionality reduced space and color the dots by the sample they belong to. 

```{r visualizing-umap-by-sample-single}

var <- "sample"
rownames(mat_red) <- spe$sample
meta <- colData(spe)[match(rownames(mat_red), spe$sample),var]

samples <- unique(meta)

bin_col <- setNames(c(adjustcolor("gray90",alpha.f=1), "palegreen4"), c(FALSE,TRUE))

plot_list <- list()

for(i in 1: length(samples)){

  sample_binary <- meta == samples[i] 
  
  background <- as.data.frame(mat_red[!sample_binary,])
  foreground <- as.data.frame(mat_red[sample_binary,])
  plot_list[[i]] <- ggplot(background, aes(x=V1, y=V2) ) +
    geom_point(color="gray90", size=0.05) + 
    geom_point(data=foreground, color="palegreen4", show.legend=F, size=0.05, alpha=0.25) +
    ggtitle(samples[i]) +
    theme(plot.title = element_text(size = 5), 
          axis.title = element_text(size=5), 
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(), 
          panel.background = element_blank(), 
          axis.line = element_line(colour = "black"))
}


plot_sublists <- split(plot_list, ceiling(seq_along(plot_list) / 33)) 

save_tiff <- function(sublist, j){
  plot <- plot_grid(plotlist = sublist, nrow=11, ncol=3) 
  ggsave(filename=file.path(path, paste("/results/Manuscript/1_Figure/sample_", j, ".tiff", sep="")), width = 7, height = 20, units = "in", device='tiff', dpi=500)
}

mapply(save_tiff, plot_sublists, 1:length(plot_sublists))
```

Next, we will visualize the cluster per sample and the sample per cluster as described [here](https://r-graph-gallery.com/48-grouped-barplot-with-ggplot2.html).
```{r barplot-cluster-per-sample, fig.width=10}
 # library
library(ggplot2)
library(tidyverse)

col_clusters <- col_clusters[order(as.integer(names(col_clusters)))]

colnames(mat) <- spe$sample
data <- data.frame(sample=colnames(mat), clusters)
data$freq <- 1

data_wide <- data %>% pivot_wider(names_from = c("clusters"), values_from = "freq", values_fn = sum, values_fill = 0)

#write.csv(data_wide, file.path(path, "results/Manuscript/1_Figure/clusters-per-sample.csv"))

data_long <- data_wide[c(1:145),] %>% pivot_longer(cols = !sample, names_to = "cluster", values_to = "count") 

data_long$cluster <- factor(data_long$cluster, levels=levels(clusters))

 
# Stacked + percent
ggplot(data_long, aes(fill=cluster, y=count, x=sample)) + 
  ggtitle("Clusters per PT sample") +
  geom_bar(position="fill", stat="identity") + 
  scale_fill_manual(values = col_clusters) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) 
```


```{r barplot-samples-per-cluster, fig.width=10}
qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]
col_sample = setNames(sample(col_vector_433, length(unique(data$sample))), unique(data$sample))

data_wide <- data %>% pivot_wider(names_from = c("sample"), values_from = "freq", values_fn = sum, values_fill = 0)

write.csv(data_wide, file.path(path, "results/Manuscript/1_Figure/samples-per-cluster.csv"))

data_long <- data_wide %>% pivot_longer(cols = !clusters, names_to = "sample", values_to = "count") 

 
# Stacked + percent
ggplot(data_long, aes(fill=sample, y=count, x=clusters)) + 
  ggtitle("Samples per cluster") +
  geom_bar(position="fill", stat="identity", show.legend = F) + 
  scale_fill_manual(values = col_sample) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) 
```
We will also save the number of cells per cluster and image. 
```{r barplot-samples-per-cluster, fig.width=20}

data <- data.frame(sample=spe$sample_id, clusters)
data$freq <- 1

data_wide <- data %>% pivot_wider(names_from = c("sample"), values_from = "freq", values_fn = sum, values_fill = 0)

write.csv(data_wide, file.path(path, "results/Manuscript/1_Figure/image-per-cluster.csv"))

```


We will now visualize the tissues per cluster. 

```{r barplot-tissues-per-cluster, fig.width=17}

colnames(mat) <- spe$sample
qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]
#col_sample = setNames(sample(col_vector_433, length(unique(data$sample))), unique(data$sample))

data <- data.frame(tissue=gsub(".*_*_", "", colnames(mat)), clusters)
data[data$tissue%in%"TU",]$tissue <- "PT"
data$freq <- 1
data_wide <- data %>% pivot_wider(names_from = c("tissue"), values_from = "freq", values_fn = sum, values_fill = 0)

#write.csv(data_wide, file.path(path, "/results/Manuscript/1_Figure/tissues-per-cluster.csv"))

data_long <- data_wide %>% pivot_longer(cols = !clusters, names_to = "tissue", values_to = "count") 

 
# Stacked + percent
ggplot(data_long, aes(fill=tissue, y=count, x=clusters)) + 
  ggtitle("Samples per cluster") +
  geom_bar(position="fill", stat="identity", show.legend = T) + 
  scale_fill_manual(values = col_tissue) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) 
```

We will now visualize the clusters in a few samples to see if they make sense.

```{r visualize-clusters-in-image, message=FALSE, fig.width=15}
library(cytomapper)
library(stringr)

tissue = "BM"

files <- list.files(file.path(path, "masks_temp", tissue))

for(i in 1: length(files)){
#for(i in 1: 1){
  
  masks <- loadImages(file.path(path, "masks_temp", tissue, files[i]), as.is = TRUE)
  
  # delete later
  names(masks) <- str_match(files[i], "^.*_.*_(.*_.*_.*_.*_.*).tif")[,2]
  c <- str_match(files[i], "^(.*)_.*_.*_.*_.*_.*_.*")[,2]
  
  mcols(masks) <- DataFrame(sample_id = names(masks))
  
  
  # plotCells(masks,
  #           object = spe,
  #           cell_id = "ObjectNumber", img_id = "sample_id",
  #           colour_by = "pg_clusters",
  #           colour = list(pg_clusters = col_clusters),
  #           display = "single",
  #           save_plot = list(filename = file.path(path, "results/Manuscript/1_Figure/", tissue, paste0("all_", files[i], "f"))),
  #           legend = NULL)


  col_celltype <- setNames(c("red", "white"), c('yes', 'no'))
  spe$binary <- 'no'
  spe[,spe$pg_clusters == as.integer(c)]$binary <- 'yes'


  plotCells(masks,
            object = spe,
            cell_id = "ObjectNumber", img_id = "sample_id",
            colour_by = "binary",
            colour = list(binary = col_celltype),
            display="single",
            save_plot = list(filename = file.path(path, "results/Manuscript/1_Figure/", tissue, paste0(files[i], "f"))),
            legend=NULL
            )

}

```

We will then examine the cell types in the images and annotate the clusters based on literature.


```{r visualize-clusters-in-image, message=FALSE, fig.width=15}
library(dplyr)
annotation <- read.csv(file.path(path, "results/Manuscript/1_Figure", "annotation_v9.csv"), sep=";")

#non-merged clusters
# BM_cluster_names <- setNames(paste0(annotation$cluster, "_", annotation$BM.annotation.2), annotation$cluster)
# PT_cluster_names <- setNames(paste0(annotation$cluster, "_", annotation$PT.annotation.2), annotation$cluster)
BM_cluster_names <- setNames(annotation$BM.annotation, annotation$cluster)
PT_cluster_names <- setNames(annotation$PT.annotation, annotation$cluster)

BM_celltypes <- recode(spe[,spe$tissue=="BM"]$pg_clusters, !!!BM_cluster_names)
PT_celltypes <- recode(spe[,spe$tissue=="PT"]$pg_clusters, !!!PT_cluster_names)

spe$celltype <- "na"
spe[,spe$tissue=="BM"]$celltype<- as.vector(BM_celltypes)
spe[,spe$tissue=="PT"]$celltype<- as.vector(PT_celltypes)
```
We will do the same for the metaclusters.

```{r visualize-clusters-in-image, message=FALSE, fig.width=15}

BM_metacluster_names <- setNames(annotation$BM.metacluster, annotation$cluster)
PT_metacluster_names <- setNames(annotation$PT.metacluster, annotation$cluster)

BM_metacelltypes <- recode(spe[,spe$tissue=="BM"]$pg_clusters, !!!BM_metacluster_names)
PT_metacelltypes <- recode(spe[,spe$tissue=="PT"]$pg_clusters, !!!PT_metacluster_names)

spe$metacluster <- "na"
spe[,spe$tissue=="BM"]$metacluster<- as.vector(BM_metacelltypes)
spe[,spe$tissue=="PT"]$metacluster<- as.vector(PT_metacelltypes)

```

Now, we have one cluster GD2+MPO+, which is double positive and a mixture of DTCs and granulocytes. We will try to seperate them with gaussian mixture models. 
```{r GD2+MPO+ subclustering}

library(mclust)

markers <- grepl("GD2|CD45|MPO|CD15|CD11b", rownames(spe)) & !grepl("mean-80|IF", rownames(spe))

tmp <- t(assay(spe[,spe$celltype=="GD2+MPO+"], "exprs")[markers,])

#roundness <- spe[,spe$celltype=="GD2+MPO+"]$solidity_clipped
#tmp <- cbind(tmp, roundness)
set.seed(20231201)
gmm <- Mclust(tmp, G = 2)
clusters <- gmm$classification

column_ha = HeatmapAnnotation(cluster = clusters, 
                              show_legend=T)

tmp2 <- t(assay(spe[,spe$celltype=="GD2+MPO+"], "exprs")[rowData(spe)$use_channel,])
ht<- Heatmap(t(tmp2), 
        column_title = "Single-cell heatmap",
        cluster_columns=F,
        show_row_dend = F,
        row_labels=sapply(strsplit(colnames(tmp2), "_"), "[[", 1),
        top_annotation=column_ha,
        column_order= order(unname(clusters), decreasing=F),
        show_column_dend = T,
        show_column_names=F,
        col=viridis(100)
        )
draw(ht)

tiff(filename = file.path(path, paste0("results/Manuscript/2_Ext_Figure/R/GMM_heatmap.tiff")),
     width = 170, height = 150, units = "mm", res=600)
draw(ht)
dev.off()

setEPS()
postscript(file.path(path, "results/Manuscript/2_Ext_Figure/R/GMM_heatmap.eps"), height=5.9, width=6.7)
draw(ht)
dev.off()
```

The cells in cluster 2 are most probably clustering together based on MPO and GD2 coexpression. We can see that some DTCs co-express CD56, some CD24 and some GATA3. To correctly assign these cells to the correct tumor celltype, we will use singleR. 
```{r singleR-tumor-cell-assignment, fig.width=10, fig.height=8}
library(BiocParallel)
library(SingleR)
markers <- rowData(spe)$use_channel
tumor_markers <- grepl("CD44|CD24|GD2|CD56|PRPH|Ki-67|CXCR4|GATA3|ELAVL4|LUM|CHGA", rownames(spe)) & !grepl("mean-80|IF", rownames(spe))
query <- t(tmp2[clusters==2,colnames(tmp2)%in%rownames(spe)[markers]])
ref <- assay(spe[markers,spe$metacluster=="tumor"], "exprs")
ref_clusters <- spe[markers,spe$metacluster=="tumor"]$celltype

predicted_labels <- SingleR(test = query, ref = ref, labels = ref_clusters, BPPARAM = MulticoreParam(workers=10, RNGseed=20231030))

plotScoreHeatmap(predicted_labels, annotation_colors=list(Labels = col_clusters), annotation_legend=F, treeheight_row=0) #define the colors first 

tiff(filename = file.path(path, paste0("results/Manuscript/2_Ext_Figure/R/singleR_prediction-score.tiff")),
     width = 90, height = 55, units = "mm", res=600)
plotScoreHeatmap(predicted_labels, annotation_colors=list(Labels = col_clusters), annotation_legend=F, treeheight_row=0) #define the colors first 
dev.off()

setEPS()
postscript(file.path(path, "results/Manuscript/2_Ext_Figure/R/singleR_prediction-score.eps"), height=2.1, width=3.5)
plotScoreHeatmap(predicted_labels, annotation_colors=list(Labels = col_clusters), annotation_legend=F, treeheight_row=0) #define the colors first 
dev.off()

labels <- predicted_labels$labels

#visualization of predicted clusters in single-cell and mean-per-cluster heatmap
column_ha = HeatmapAnnotation(predicted = labels,
                              show_legend=T)

ht<- Heatmap(t(tmp2[clusters==2,]), 
        column_title = "Single-cell heatmap",
        cluster_columns=F,
        top_annotation=column_ha,
        column_order= order(unname(labels), decreasing=F),
        show_column_dend = T,
        show_column_names=F,
        col=viridis(100)
        )
draw(ht)

mean <- aggregate(tmp2[clusters==2,], list(predicted_labels$labels), mean)
mean_t <- t(mean[,2:ncol(mean)])
colnames(mean_t) <- mean[,"Group.1"]
mean_scaled <- t(apply(mean_t, 1, RESCALE, to=0:1))

column_ha = HeatmapAnnotation(ncells=anno_barplot(as.data.frame(table(predicted_labels$labels))[,"Freq"], height = unit(1, "cm")), 
                              show_legend=F)

Heatmap(mean_scaled[rownames(mean_scaled)%in%rownames(spe)[tumor_markers],], 
        column_title = "Mean per cluster (protein data)",
        top_annotation = column_ha, 
        cluster_columns=T,
        show_column_dend = T,
        show_column_names=T,
        show_row_names=T,
        col=viridis(100)
        )

#Assign predicted labels
dtc <- rownames(tmp2[clusters==2,])
no_dtc <- rownames(tmp2[clusters==1,])
spe[,colnames(spe)%in%dtc]$celltype <- labels
spe[,colnames(spe)%in%no_dtc]$celltype <- "other"
spe[,colnames(spe)%in%dtc]$metacluster <- "tumor"
```

Since, the granulocytes are wrongly classified with immature and mature clusters being a mixture of both round and segmented nuclei, we will try to cluster this metacluster based on the morphological feature roundness only.
```{r granulocyte_subclustering, fig.width=15}

spe_gran <- spe[,spe$celltype=="granulocyte" & spe$tissue=="BM"]

col_celltype_gran <- setNames(c("#a6611a", "#dfc27d", "#80cdc1", "#018571"), c('neutrophil', 'band_cell', 'myelocyte', 'myeloblast'))
spe_gran$gran_type <- 'neutrophil'
spe_gran[,spe_gran$solidity>=0.83 & spe_gran$solidity<0.93]$gran_type <- 'band cell'
spe_gran[,spe_gran$solidity>=0.93 & spe_gran$solidity<0.97]$gran_type <- 'myelocyte'
spe_gran[,spe_gran$solidity>=0.97]$gran_type <- 'myeloblast'

files <- list.files(file.path(path, "masks_temp/BM"))
#for(i in 1: length(files)){
for(i in 1: 1){
  masks <- loadImages(file.path(path, "masks_temp/BM", files[i]), as.is = TRUE)
  
  mcols(masks) <- DataFrame(sample_id = names(masks))

  plotCells(masks,
            object = spe_gran,
            cell_id = "ObjectNumber", img_id = "sample_id",
            colour_by = "gran_type",
            colour = list(gran_type = col_celltype_gran),
            display="single",
            #save_plot = list(filename = file.path(path, "results/Manuscript/1_Figure", paste0("2", files[i], "f"))),
            #legend=NULL
            )
  
    # plotCells(masks,
    #         object = spe,
    #         cell_id = "ObjectNumber", img_id = "sample_id",
    #         colour_by = "celltype",
    #         colour = list(celltype = metadata(spe)$color_vector$col_celltype),
    #         display="single",
    #         save_plot = list(filename = file.path(path, "results/Manuscript/1_Figure", paste0("2", files[i], "f"))),
    #         legend=NULL
    #         )
}
```

We will now assign the determined granulocyte subtypes for the BM cells. 

```{r granulocyte_subclustering, fig.width=15}
#assign new granulocyte celltypes
spe[,colnames(spe_gran)]$celltype <- spe_gran$gran_type
spe[,spe$tissue=="PT" & spe$celltype=="granulocyte"]$celltype <- "neutrophil"
```

We will manually select distinct colors.

```{r colors, message=FALSE, fig.height=17}
col_clusters <- c(#"GD2+MPO+"=rgb(255,255,162, maxColorValue = 255),
                #"DTC"=rgb(255,221,57, maxColorValue = 255),
                #"DTC"=rgb(246,246,0, maxColorValue = 255),
                "other"=rgb(255,255,209, maxColorValue = 255),
                "LIN- Ki67+"=rgb(254,217,166, maxColorValue = 255),
                "HSPC"=rgb(253,180,98, maxColorValue = 255),
                #"GZMB+ S100B- DC/NK"=rgb(255,137,0, maxColorValue = 255),
                "GZMB+ S100B- DC/NK"=rgb(28,146,255, maxColorValue = 255),
                "B cell"=rgb(168,128,250, maxColorValue = 255),
                "GZMB+ S100B+ DC/NK"=rgb(0,255,255, maxColorValue = 255),
                #"HLA-DR+ CXCR4+"=rgb(152,78,163, maxColorValue = 255),
                #"B cell"=rgb(82,75,181, maxColorValue = 255),
                "CD8+ naive T cell"=rgb(188,128,189, maxColorValue = 255),
                "DN GZMB+ T cell"=rgb(190,174,212, maxColorValue = 255),
                "CD8+ S100B+ T cell"=rgb(253,218,236, maxColorValue = 255),
                "CD8+ GZMB+ PD1lo T cell"=rgb(251,154,153, maxColorValue = 255),
                "CD4+ naive T cell"=rgb(242,3,137, maxColorValue = 255),
                "DN T cell"=rgb(251,128,114, maxColorValue = 255),
                "CD4+ PD1+ T cell"=rgb(254,142,175, maxColorValue = 255),
                "CD8+ PDL1lo T cell"=rgb(240,0,240, maxColorValue = 255),
                "CD4+ S100B+ T cell"=rgb(227,26,28, maxColorValue = 255),
                "dense T cell region"=rgb(145,2,144, maxColorValue = 255),
                "CD24+ marker-lo TC"=rgb(230,245,201, maxColorValue = 255),
                "early SYM-like TC"=rgb(204,235,197, maxColorValue = 255),
                "neural progenitor"=rgb(238,118,0, maxColorValue = 255),
                "bridge-like TC"=rgb(198,255,72, maxColorValue = 255),
                "GATA3hi TC"=rgb(65,208,57, maxColorValue = 255),
                "CD24- marker-lo TC"=rgb(60,119,97, maxColorValue = 255),
                "CD14+ PDL1- MO"=rgb(104,180,238, maxColorValue = 255),
                "monoblast-like"=rgb(128,177,211, maxColorValue = 255),
                "CD14+ PDL1+ MO"=rgb(51,180,180, maxColorValue = 255),
                "CHGAhi TC"=rgb(102,166,30, maxColorValue = 255),
                "Ki67hi TC"=rgb(149,144,89, maxColorValue = 255),
                "CXCR4hi TC"=rgb(210,205,126, maxColorValue = 255),
                "CD44+ mes-like TC"=rgb(27,158,119, maxColorValue = 255),
                "GD2lo TC"=rgb(102,194,165, maxColorValue = 255),
                "CD14- PDL1- MO/DC"=rgb(39,130,187, maxColorValue = 255),
                "CD14- PDL1+ MO/DC"=rgb(166,206,227, maxColorValue = 255),
                "pDC"=rgb(212,241,240, maxColorValue = 255),
                "band cell"=rgb(169,189,210, maxColorValue = 255),
                "myelocyte"=rgb(217,217,217, maxColorValue = 255),          
                #"granulocyte"=rgb(251,241,232, maxColorValue = 255),
                "myeloblast"=rgb(2,0,138, maxColorValue = 255),
                "neutrophil"=rgb(113,113,113, maxColorValue = 255),
                "fibroblast/endothel"=rgb(212,89,67, maxColorValue = 255),
                "schwann cell"=rgb(156,76,24, maxColorValue = 255),
                "MSC"=rgb(174,129,37, maxColorValue = 255),
                "HPC"=rgb(212,143,75, maxColorValue = 255),
                "Ki67+ CXCR4+ cell"=rgb(254,215,0, maxColorValue = 255))

pie(rep(1, length(col_clusters)), col=col_clusters, labels=names(col_clusters))

metadata(spe)$color_vectors$col_celltype <- col_clusters
```

```{r colors, message=FALSE, fig.height=17}
col_metacluster <- c("tumor"="palegreen4",
                  #"T-cell"="#FB8072",
                  "T cell"="indianred",
                  "granulocyte"="#80B1D3",
                  "MO/DC/NK"="#8DD3C7",
                  #"dendritic"="paleturquoise2",
                  "B cell"="#BC80BD",
                  "other"="goldenrod1",
                  "progenitor"="tan2",
                  "mesenchymal"="tan4"
                  #"LIN- prol."="tan"
                )
                
pie(rep(1, length(col_metacluster)), col=col_metacluster, labels=names(col_metacluster))


metadata(spe)$color_vectors$col_metacluster <- col_metacluster
```


## Save objects

Finally, the spe object can be saved for further downstream 
processing and analysis.

```{r save-objects-read-data}

saveRDS(spe, file.path(path, "results/Manuscript/spe_clustered.rds"))
```
