==> code/R/20231214_Manuscript/1_Figure/09_cohort_overview.Rmd <==
---
title: "Visualize metadata"
output: html_document
date: '2022-08-15'
---

```{r setup, include=FALSE, cache=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "120%")
```

# Visualize metadata with complex heatmap

Complex Heatmap allows visualization of complex heatmap as explained in these [videos](https://www.youtube.com/watch?v=NlMga8pK2Pw) and [notebooks](https://github.com/brandonyph/ComplexHeatmap_Walkthrough). We will try to show the cohort metadata together with the the heatmaps generated in clustering.RMD. 

## Load data

We will first read in the previously generated `SpatialExperiment` object.

```{r read-data-pheno, message=FALSE}
library(SpatialExperiment)
library(ComplexHeatmap)
library(RColorBrewer)
library(dittoSeq)
library(paletteer)
library(bruceR)
library(circlize)
library(viridis)

path <- "/mnt/Multimodal_Imaging_Daria/PIPELINE/NEWEST_RESULTS"

spe <- readRDS(file.path(path,"results/Manuscript/spe_clustered.rds"))

```


First, we will define the color code for the metadata. Find all built-in R color names [here](https://www.farb-tabelle.de/de/farbtabelle.htm). The [paletteer package](https://r-charts.com/color-palettes/) offers a wide range of color palettes.  

```{r define-color-code, message=FALSE}

meta_path <- file.path(path, '20231128_metadata_complete.csv')

qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector_74 = unlist(mapply(brewer.pal, qual_col_pals$maxcolors, rownames(qual_col_pals)))

col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]

meta_all = read.csv(meta_path)
meta_all <- meta_all[order(meta_all$Study_ID),] #order by study-id

exclude <- replace(x = meta_all$exclude, list =  !meta_all$exclude %in% c('no', 'maybe'), values =  'partially')

DE_year <- sub("-.*", "", meta_all$DE_date)

col_sample = setNames(sample(col_vector_433, length(unique(meta_all$Sample))), unique(meta_all$Sample))
col_puncture_side = setNames(brewer.pal(length(unique(meta_all$Puncture_side)), 'Set3'), unique(meta_all$Puncture_side))
col_staining_batch = setNames(sample(col_vector_74, length(unique(meta_all$Staining_date))), unique(meta_all$Staining_date))
col_stainer = setNames(c('#a6611a', '#dfc27d', '#80cdc1', '#018571'), unique(meta_all$Stainer))
col_imc_batch = setNames(sample(col_vector_74, length(unique(meta_all$IMC_date))), unique(meta_all$IMC_date))
col_imc_score = colorRamp2(c(min(meta_all$IMC_score, na.rm = TRUE), max(meta_all$IMC_score, na.rm = TRUE)/2, max(meta_all$IMC_score, na.rm = TRUE)), c("red", "white", "blue")) 
col_study_id = setNames(sample(col_vector_433, length(unique(meta_all$Study_ID))), unique(meta_all$Study_ID))
col_DE_year = setNames(sample(col_vector_74, length(unique(DE_year))), unique(DE_year))
col_timepoint = setNames(c('#7fc97f', '#beaed4', '#fdc086' , '#ffff99', '#386cb0'), unique(meta_all$Timepoint))
col_aberrations = setNames(c('#a6611a','#018571','#80cdc1','#f5f5f5'), c(0,1,2,9))
col_stages = setNames(c('#a6cee3', '#1f78b4', '#b2df8a', '#33a02c', '#fb9a99', '#e31a1c'), unique(meta_all$stage4))
col_efs_dur = colorRamp2(c(min(meta_all$efs_dur, na.rm = TRUE), max(meta_all$efs_dur, na.rm = TRUE)/2, max(meta_all$efs_dur, na.rm = TRUE)), paletteer_c("ggthemes::Red-Green Diverging", 3))
col_tissue = setNames(c('aquamarine4','brown'), unique(meta_all$Tissue))
col_control = setNames(c('mediumseagreen', 'gray77'), unique(meta_all$control))
col_exclude = setNames(c('#018571','#a6611a','#80cdc1'), unique(exclude))
col_progress = setNames(c('gray77', 'palegreen3', 'firebrick1'), c('n.a.', 0, 1))
```

Now, we will plot an overview of the cohort.

```{r overview-cohort, message=FALSE, fig.height=25}
columnAnnotation(sample=meta_all$Sample, col=list(sample=col_sample), annotation_legend_param = list(sample = list(grid_height=unit(0.05, "cm"), labels_gp = gpar(fontsize = 5)))) %v%
  columnAnnotation(exclusion=exclude, col=list(exclusion=col_exclude)) %v%
  columnAnnotation(imc_score=meta_all$IMC_score, col=list(imc_score=col_imc_score))  %v%
  columnAnnotation(staining_batch=meta_all$Staining_date, col=list(staining_batch=col_staining_batch)) %v%
  columnAnnotation(stainer=meta_all$Stainer, col=list(stainer=col_stainer)) %v%
  columnAnnotation(imc_batch=meta_all$IMC_date, col=list(imc_batch=col_imc_batch)) %v%
  columnAnnotation(DE_year=meta_all$Staining_date, col=list(staining_batch=col_staining_batch))  %v%
  columnAnnotation(control=meta_all$control, col=list(control=col_control)) %v%
  columnAnnotation(study_id=meta_all$Study_ID, col=list(study_id=col_study_id))  %v%
  columnAnnotation(timepoint=meta_all$Timepoint, col=list(timepoint=col_timepoint))  %v%
  columnAnnotation(tissue=meta_all$Tissue, col=list(tissue=col_tissue)) %v%
  columnAnnotation(stage=meta_all$stage4, col=list(stage=col_stages)) %v%
  columnAnnotation(progression=meta_all$efs, col=list(progression=col_progress))  %v%
  columnAnnotation(efs_dur=meta_all$efs_dur, col=list(efs_dur=col_efs_dur)) %v%
  columnAnnotation(MYCN_amp=meta_all$MYCN_amp, col=list(MYCN_amp=col_aberrations)) %v%#, annotation_legend_param = list(MYCN_amp = list(title='aberration', labels = c('no', 'yes', 'het', 'NA'))))  %v%
  columnAnnotation(ALK_amp=meta_all$ALK_amp, col=list(ALK_amp=col_aberrations), show_legend=FALSE) %v%
  columnAnnotation(ALK_mut=meta_all$ALK_mut, col=list(ALK_mut=col_aberrations), show_legend=FALSE) %v%
  columnAnnotation(ATRX_del=meta_all$ATRX_del, col=list(ATRX_del=col_aberrations), show_legend=FALSE) %v%
  columnAnnotation(TERT_rear=meta_all$TERT_rear, col=list(TERT_rear=col_aberrations), show_legend=FALSE) %v%
  columnAnnotation(X1ploss=meta_all$X1p_loss, col=list(X1ploss=col_aberrations), show_legend=FALSE) %v%
  columnAnnotation(X1qgain=meta_all$X1q_gain, col=list(X1qgain=col_aberrations), show_legend=FALSE) %v%
  columnAnnotation(X11qloss=meta_all$X11q_loss, col=list(X11qloss=col_aberrations), show_legend=FALSE
)
```
Now we will visualize the available samples per patient in a tidier format.

```{r meta_data-timepoint, message=FALSE, fig.height=17, fig.width=4}
meta_positive <-meta_all[meta_all$control == 'no',]
meta_pos_reduced <- meta_positive[c("Study_ID", "Tissue", "Timepoint")]
meta_pos_reduced$new <- 1 
meta_pos_reduced_tidy <- meta_pos_reduced %>% pivot_wider(names_from = c("Tissue", "Timepoint"), values_from = "new")
meta_timepoint <- meta_pos_reduced_tidy[, c("Study_ID", "PT_DE", "BM_DE", "BM_RE1", "BM_RE2", "BM_REL")]
meta_timepoint[is.na(meta_timepoint)] <- 0
#meta_timepoint <- meta_timepoint %>% rowwise() %>% mutate(n_samples = sum(c(PT_DE, BM_DE, BM_RE1, BM_RE2, BM_REL)))
meta_timepoint <- as.data.frame(meta_timepoint)
rownames(meta_timepoint) <- meta_timepoint$Study_ID
meta_timepoint$Study_ID <- NULL

type <- c()

for(i in 1:nrow(meta_timepoint)) {
  if (meta_timepoint[i,'PT_DE']==1 && meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_RE1']==1 || meta_timepoint[i,'BM_RE2']==1 || meta_timepoint[i,'BM_REL']==1)){
    type <- c(type, "spatial and temporal")
  } else if (meta_timepoint[i,'PT_DE']==1 && meta_timepoint[i,'BM_DE']==1){
    type <- c(type, "spatial")
  } else if (meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_RE1']==1 || meta_timepoint[i,'BM_RE2']==1 || meta_timepoint[i,'BM_REL']==1)){
    type <- c(type, "temporal")
  } else if (meta_timepoint[i,'PT_DE']==1 && meta_timepoint[i,'BM_DE']==0 && meta_timepoint[i,'BM_RE1']==0 && meta_timepoint[i,'BM_RE2']==0 && meta_timepoint[i,'BM_REL']==0){
    type <- c(type, "PT_only") 
  } else if (meta_timepoint[i,'PT_DE']==0 && meta_timepoint[i,'BM_DE']==1 && meta_timepoint[i,'BM_RE1']==0 && meta_timepoint[i,'BM_RE2']==0 && meta_timepoint[i,'BM_REL']==0){
    type <- c(type, "BM_DE_only") 
  } else if (meta_timepoint[i,'PT_DE']==0 && meta_timepoint[i,'BM_DE']==0 && meta_timepoint[i,'BM_RE1']==1 && meta_timepoint[i,'BM_RE2']==0 && meta_timepoint[i,'BM_REL']==0){
    type <- c(type, "BM_RE1_only")
  } else if (meta_timepoint[i,'PT_DE']==0 && meta_timepoint[i,'BM_DE']==0 && meta_timepoint[i,'BM_RE1']==0 && meta_timepoint[i,'BM_RE2']==1 && meta_timepoint[i,'BM_REL']==0){
    type <- c(type, "BM_RE2_only") 
  } else if (meta_timepoint[i,'PT_DE']==0 && meta_timepoint[i,'BM_DE']==0 && meta_timepoint[i,'BM_RE1']==0 && meta_timepoint[i,'BM_RE2']==0 && meta_timepoint[i,'BM_REL']==1){
    type <- c(type, "BM_REL_only") 
  } else {
    type <- c(type, "other")
  }
}

col_n = setNames(c('gray50','#fcae91'), c(0,1))
col_type = setNames(c('#a6cee3','#1f78b4','#b2df8a','#33a02c','#fb9a99','#e31a1c','#fdbf6f','#ff7f00','#cab2d6'), unique(type))

order <- c("PT_only", "BM_DE_only", "BM_RE1_only", "BM_RE2_only", "BM_REL_only", "spatial and temporal", "spatial", "temporal", "other")
type <- factor(type, levels=rev(order))

Heatmap(meta_timepoint, 
        #cluster_rows=FALSE,
        show_row_dend = FALSE,
        cluster_columns=FALSE, 
        col=col_n,
        rect_gp = gpar(col = "white", lwd = 4), 
        row_names_gp = gpar(fontsize = 8), 
        column_names_gp = gpar(fontsize = 8), 
        split=type,
        cluster_row_slices=F,
        row_title=NULL,
        heatmap_legend_param = list(title="available"),
        top_annotation=columnAnnotation(total = anno_barplot(colSums(meta_timepoint), 
                                                             add_numbers=TRUE,
                                                             axis=FALSE,
                                                             numbers_offset = unit(1, "mm"),
                                                             annotation_name_gp= gpar(fontsize = 8))),
        left_annotation=rowAnnotation(foo = anno_block(gp = gpar(fill = col_type[rev(order)]),
                                                       labels = summary(type), 
                                                       labels_gp = gpar(col = "white", fontsize = 10))
                                      ))

lgd = Legend(at = order, title = "analysis", legend_gp = gpar(fill = col_type[order]))

draw(lgd, test = "add labels and title")

```

```{r meta_data-timepoint, message=FALSE, fig.height=17, fig.width=8}
meta_positive <-meta_all[meta_all$control == 'no',]
meta_pos_reduced <- meta_positive[c("Study_ID", "Tissue", "Timepoint", "MYCN_amp", "ALK_amp", "ALK_mut", "ATRX_del", "TERT_rear", "X1p_loss", "X1q_gain", "X11q_loss", "efs", "age_at_diag")]
meta_pos_reduced$new <- 1 
meta_pos_reduced_tidy <- meta_pos_reduced %>% pivot_wider(names_from = c("Tissue", "Timepoint"), values_from = "new")
meta_timepoint <- meta_pos_reduced_tidy[, c("Study_ID", "PT_DE", "BM_DE", "BM_RE1", "BM_RE2", "BM_REL", "MYCN_amp", "ALK_amp", "ALK_mut", "ATRX_del", "TERT_rear", "X1p_loss", "X1q_gain", "X11q_loss", "efs", "age_at_diag")]
meta_timepoint[is.na(meta_timepoint)] <- 0
meta_timepoint <- as.data.frame(meta_timepoint)
rownames(meta_timepoint) <- meta_timepoint$Study_ID
meta_timepoint$Study_ID <- NULL

type <- c()

for(i in 1:nrow(meta_timepoint)) {
  if (meta_timepoint[i,'PT_DE']==1 && meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_RE1']==1 && meta_timepoint[i,'BM_RE2']==1 && meta_timepoint[i,'BM_REL']==1)){
    type <- c(type, "spatial and temporal1")
  } else if (meta_timepoint[i,'PT_DE']==1 && meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_RE1']==1 && meta_timepoint[i,'BM_RE2']==1)){
    type <- c(type, "spatial and temporal2")
  } else if (meta_timepoint[i,'PT_DE']==1 && meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_RE1']==1)){
    type <- c(type, "spatial and temporal3")
  } else if (meta_timepoint[i,'PT_DE']==1 && meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_RE2']==1)){
    type <- c(type, "spatial and temporal4")
  } else if (meta_timepoint[i,'PT_DE']==1 && meta_timepoint[i,'BM_DE']==1){
    type <- c(type, "spatial")
  } else if (meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_RE1']==1 && meta_timepoint[i,'BM_RE2']==1 && meta_timepoint[i,'BM_REL']==1)){
    type <- c(type, "temporal1")
  } else if (meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_RE1']==1 && meta_timepoint[i,'BM_RE2']==1)){
    type <- c(type, "temporal2")
  } else if (meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_RE1']==1)){
    type <- c(type, "temporal3")
  } else if (meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_RE2']==1)){
    type <- c(type, "temporal4")
  } else if (meta_timepoint[i,'BM_DE']==1 && (meta_timepoint[i,'BM_REL']==1)){
    type <- c(type, "temporal5")
  } else if (meta_timepoint[i,'PT_DE']==1 && meta_timepoint[i,'BM_DE']==0 && meta_timepoint[i,'BM_RE1']==0 && meta_timepoint[i,'BM_RE2']==0 && meta_timepoint[i,'BM_REL']==0){
    type <- c(type, "PT_only") 
  } else if (meta_timepoint[i,'PT_DE']==0 && meta_timepoint[i,'BM_DE']==1 && meta_timepoint[i,'BM_RE1']==0 && meta_timepoint[i,'BM_RE2']==0 && meta_timepoint[i,'BM_REL']==0){
    type <- c(type, "BM_DE_only") 
  } else if (meta_timepoint[i,'PT_DE']==0 && meta_timepoint[i,'BM_DE']==0 && meta_timepoint[i,'BM_RE1']==1 && meta_timepoint[i,'BM_RE2']==0 && meta_timepoint[i,'BM_REL']==0){
    type <- c(type, "BM_RE1_only")
  } else if (meta_timepoint[i,'PT_DE']==0 && meta_timepoint[i,'BM_DE']==0 && meta_timepoint[i,'BM_RE1']==0 && meta_timepoint[i,'BM_RE2']==1 && meta_timepoint[i,'BM_REL']==0){
    type <- c(type, "BM_RE2_only") 
  } else if (meta_timepoint[i,'PT_DE']==0 && meta_timepoint[i,'BM_DE']==0 && meta_timepoint[i,'BM_RE1']==0 && meta_timepoint[i,'BM_RE2']==0 && meta_timepoint[i,'BM_REL']==1){
    type <- c(type, "BM_REL_only") 
  } else {
    type <- c(type, "other")
  }
}

order <- c("spatial and temporal1", "spatial and temporal2", "spatial and temporal3", "spatial and temporal4", "spatial", "temporal1", "temporal2", "temporal3", "temporal4","temporal5","other", "PT_only", "BM_DE_only", "BM_RE1_only", "BM_RE2_only", "BM_REL_only")
type <- factor(type, levels=rev(order))

column_ha = rowAnnotation(MYCN_amp = meta_timepoint$MYCN_amp,
                          X1p_loss=meta_timepoint$X1p_loss,
                          progression=meta_timepoint$efs,
                          age=anno_barplot(meta_timepoint$age_at_diag, height = unit(3, "cm")),
                              col=list(MYCN_amp=col_aberrations,
                                       X1p_loss=col_aberrations,
                                       progression=col_progress
                                       ),
                              show_legend=T)

setEPS()
postscript(file.path(path, "results/Manuscript/1_Figure/R/cohort_overview.eps"), height=18, width=5)

Heatmap(meta_timepoint[, c("PT_DE", "BM_DE", "BM_RE1", "BM_RE2", "BM_REL")], 
        show_row_dend = FALSE,
        cluster_columns=FALSE, 
        col=col_n,
        rect_gp = gpar(col = "white", lwd = 4), 
        row_names_gp = gpar(fontsize = 8), 
        column_names_gp = gpar(fontsize = 8), 
        row_order=order(type, decreasing=F),
        cluster_row_slices=F,
        row_title=NULL,
        left_annotation = column_ha, 
        heatmap_legend_param = list(title="available"))

dev.off()
```

## Session Info

<details>
   <summary>SessionInfo</summary>
   
```{r, echo = FALSE}
sessionInfo()
```
</details>
